% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{albumen}
\alias{albumen}
\title{Variación en el color del albumen de semillas de plantas de Mendel}
\format{
A data frame with 10 rows and 2 columns:
\itemize{
\item \code{amarillo}: cantidad de semillas con albumen amarillo obtenidas por planta
\item \code{verde}: cantidad de semillas con albumen verde obtenidas por planta
}
}
\usage{
data(albumen)
}
\description{
Resultados de variación de planta a planta en el color del albumen (amarillo o verde) en diez plantas de un estudio de Mendel.
}
\details{
Mendel también observó la variación de planta a planta en un experimento sobre el color del albumen de las semillas. Los resultados para diez plantas seleccionadas se registran en este set de datos.
}
\keyword{internal}
