% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{alimento}
\alias{alimento}
\title{Contenido vitamínico en alimento animal}
\format{
A data frame with 16 rows and 1 column:
\itemize{
\item \code{vitaminas}: contenido de vitaminas por cada 100 g
}
}
\usage{
data(alimento)
}
\description{
Datos del contenido de vitaminas (por cada 100 g) en una muestra de 16 animales alimentados con cierto alimento.
}
\details{
Un nutricionista animal desea estimar el contenido vitamínico de cierto alimento. Toma una muestra de 16 animales y registra los contenidos de vitaminas por cada 100 g.
}
\keyword{internal}
