% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{arvejas}
\alias{arvejas}
\title{Frecuencias fenotípicas en arvejas de Mendel}
\format{
A data frame with 4 rows and 2 columns:
\itemize{
\item \code{freq}: cantidad de arvejas observadas (frecuencia)
\item \code{fenotipo}: fenotipo mendeliano observado ("LA", "LV", "RA", "RV")
}
}
\usage{
data(arvejas)
}
\description{
Frecuencias observadas de textura y color en 556 arvejas clasificadas según fenotipo mendeliano.
}
\details{
En una variedad de arvejas, la textura y color de la semilla se rigen por caracteres mendelianos, siendo "lisas" (L) y "amarillas" (A) fenotipos dominantes, mientras que "rugosas" (R) y "verdes" (V) son recesivos. La proporción teórica del número de arvejas LA, LV, RA y RV es 9:3:3:1. Se clasificaron 556 arvejas según su fenotipo y se obtuvieron los resultados registrados.
}
\keyword{internal}
