% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{balanceado_cerdos}
\alias{balanceado_cerdos}
\title{Aumento de peso en cerdos con alimento balanceado}
\format{
A data frame with 20 rows and 1 column:
\itemize{
\item \code{aumento}: aumento de peso en gramos
}
}
\usage{
data(balanceado_cerdos)
}
\description{
Datos de aumentos de peso (en gramos) de una muestra de 20 cerdos tomados al azar para evaluar la afirmación de un fabricante sobre su alimento balanceado.
}
\details{
Un fabricante dedicado a la elaboración de alimento balanceado para cerdos afirma que su producto aumenta el peso promedio en 200 gr. En una muestra de 20 cerdos tomados al azar se obtuvieron los siguientes aumentos de peso (en gramos).
}
\keyword{internal}
