% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{barrenador}
\alias{barrenador}
\title{Eficacia de insecticida para control de barrenador del brote de soja}
\format{
A data frame with 2050 rows and 3 columns:
\itemize{
\item \code{Muestra}: identificador numérico de la planta (muestra)
\item \code{Control}: si la plaga fue controlada ("si", "no")
\item \code{Lote}: identificador del lote experimental (1 o 2)
}
}
\usage{
data(barrenador)
}
\description{
Datos sobre el control del barrenador del brote de soja (\emph{Epinotia aporema}) tras la aplicación de un insecticida en dos lotes de cultivo.
}
\details{
Datos correspondientes a un experimento realizado para evaluar la eficacia de un insecticida destinado al control del barrenador del brote de soja (\emph{Epinotia aporema}). Se seleccionaron plantas afectadas por la plaga en dos lotes experimentales diferentes (Lote 1 y Lote 2) y se les aplicó la dosis de producto recomendada, registrando de forma binaria si la plaga fue controlada eficazmente en cada planta.
}
\keyword{internal}
