% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{caballos}
\alias{caballos}
\title{Medicamentos en caballos de carrera}
\format{
A data frame with 12 rows and 2 columns:
\itemize{
\item \code{TR}: tipo de medicamento suministrado.
\item \code{TI}: tiempo registrado en segundos.
}
}
\source{
--
}
\usage{
data(caballos)
}
\description{
Respuesta de caballos de carrera a 3 medicamentos alternativos.
}
\details{
Para el experimento se utilizaron 12 caballos de características similares
(edad, velocidad, sanidad). Los medicamentos evaluados se suministraron por
triplicado de manera individual dejando un grupo de animales testigo sin medicar.

La respuesta se midió registrando los tiempos (TI) en segundos empleados en
una corrida de 300 metros en pista normal.
}
\keyword{internal}
