% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{camarones}
\alias{camarones}
\title{Efecto de temperatura, densidad y salinidad en el crecimiento de camarones}
\format{
A data frame with 36 rows and 5 columns:
\itemize{
\item \code{T}: factor que identifica la temperatura del agua (25 o 35 °C).
\item \code{D}: factor que identifica la densidad de población (80 o 160 individuos).
\item \code{S}: factor que identifica la salinidad del agua (10, 25 o 40 g/L).
\item \code{rep}: número de réplica o repetición (1 a 3).
\item \code{Y}: tasa de crecimiento registrada.
}
}
\source{
--
}
\usage{
data(camarones)
}
\description{
Estudio llevado a cabo para determinar cómo la densidad de población del
estanque (D), la temperatura (T) y la salinidad del agua (S) influyen sobre
la tasa de crecimiento de camarones criados en acuarios.
}
\details{
Se evalúa si la densidad de población, la temperatura y la salinidad influyen
de manera conjunta (interacciones) o independiente sobre la tasa de
crecimiento de los camarones criados en condiciones controladas de acuarios.
}
\keyword{internal}
