% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{cebada_salinidad}
\alias{cebada_salinidad}
\title{Efecto de salinidad y desarrollo en el consumo de agua en cebada}
\format{
A data frame with 18 rows and 5 columns:
\itemize{
\item \code{S}: salinidad del sustrato (0, 6 o 12 bares).
\item \code{D}: estado de desarrollo de la planta de cebada (14, 21 o 28 días).
\item \code{C}: consumo de agua registrado.
\item \code{D_ord}: estado de desarrollo de la planta transformado a factor ordenado (14 < 21 < 28).
\item \code{S_ord}: salinidad del sustrato transformada a factor ordenado (0 < 6 < 12).
}
}
\source{
--
}
\usage{
data(cebada_salinidad)
}
\description{
Ensayo para estudiar el efecto de la salinidad del sustrato S (0, 6 y 12
bares) y el estado de desarrollo de la planta de cebada D (14, 21 y 28 días)
en el consumo de agua C.
}
\details{
Los 9 tratamientos resultantes de la combinación de los factores de
salinidad y días de desarrollo fueron asignados al azar a las unidades
experimentales siguiendo un diseño completamente al azar (DCA).
}
\keyword{internal}
