% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkminiCRAN.R
\name{check_mirrors}
\alias{check_mirrors}
\title{Chequear lista de mirrors}
\usage{
check_mirrors(mirrors)
}
\arguments{
\item{mirrors}{Espera un vector de caracteres con los diferentes mirrors}
}
\value{
Devuelve una url de un mirror valida
}
\description{
Función que chequea una lista de mirrors y elige el optimo
}
\details{
Recorre la lista de mirrors en orden y chequea que funciona.
Si funciona lo retorna.
Si no funciona pasa al siguiente, si ninguno funciona devuelve error
}
