% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.test.R
\name{chi_values}
\alias{chi_values}
\title{Funcion interna para calcular p-valor e intervalo de confianza para Chi-cuadrado}
\usage{
chi_values(
  chi_calc,
  df,
  ss,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95
)
}
\arguments{
\item{chi_calc}{Numerico. Estadistico Chi-cuadrado calculado.}

\item{df}{Numerico. Grados de libertad (n - 1).}

\item{ss}{Numerico. Suma de cuadrados (df * s2).}

\item{alternative}{String de caracteres que especifica la hipotesis alternativa. Debe ser \code{"two.sided"} (default), \code{"greater"} o \code{"less"}.}

\item{conf.level}{Numerico. Nivel de confianza del intervalo. Por defecto es 0.95.}
}
\value{
Un listado con \code{p_val} (valor p) y \code{conf_int} (intervalo de confianza).
}
\description{
Funcion auxiliar para calcular p-valor e intervalo de confianza para pruebas de una varianza bajo distribucion Chi-cuadrado.
}
\keyword{internal}
