% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{citricos}
\alias{citricos}
\title{Rendimiento de fruta en cítricos bajo diferentes pesticidas}
\format{
A data frame with 24 rows and 4 columns:
\itemize{
\item \code{Variedad}: factor que identifica la variedad de cítrico (1, 2, 3).
\item \code{Pesticida}: factor que identifica el tipo de pesticida (1, 2, 3, 4).
\item \code{rep}: número de réplica o repetición (1, 2).
\item \code{rend}: rendimiento en bushels por árbol.
}
}
\source{
--
}
\usage{
data(citricos)
}
\description{
Experimento para determinar los efectos de 4 pesticidas en el rendimiento de
fruta de 3 variedades de cítricos.
}
\details{
Se seleccionaron 8 árboles de cada variedad de cítrico y se asignó al azar uno
de los pesticidas a dos árboles de cada variedad (réplicas). Las aplicaciones
se realizaron siguiendo las dosis de marbete y se midió el rendimiento en
bushels por árbol.

Un bushel equivale aproximadamente a 18.14 kg.
}
\keyword{internal}
