% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{corderos}
\alias{corderos}
\title{Peso de corderos}
\format{
A data frame with 20 rows and 1 column:
\itemize{
\item \code{peso}: peso en kilogramos
}
}
\usage{
data(corderos)
}
\description{
Datos de pesos en kilogramos de una muestra de 20 corderos.
}
\details{
Una muestra de 20 corderos indica los pesos en kg.
}
\keyword{internal}
