% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_groups.R
\docType{data}
\name{datasets_estb}
\alias{datasets_estb}
\title{Datasets para Estadística Básica}
\description{
Lista de datasets utilizados en la asignatura Estadística Básica.
}
\section{Datasets}{

\itemize{
\item \link{actividad}
\item \link{agua_forrajera}
\item \link{albumen}
\item \link{algodon}
\item \link{alimento}
\item \link{area_trigo}
\item \link{arvejas}
\item \link{balanceado_cerdos}
\item \link{barrenador}
\item \link{calificaciones}
\item \link{capturados}
\item \link{cerdos}
\item \link{coleoptero}
\item \link{colesterol}
\item \link{concentracion_insecticida}
\item \link{corderos}
\item \link{dieta_bufalos}
\item \link{dietas_pollos}
\item \link{dosis_tomate}
\item \link{educacion}
\item \link{efectividad_sembradora}
\item \link{especies}
\item \link{establecimientos}
\item \link{fungicida}
\item \link{ganancia_novillos}
\item \link{germinacion}
\item \link{glifosato}
\item \link{hileras_maiz}
\item \link{huevos}
\item \link{insectos}
\item \link{lesiones_hojas}
\item \link{lisina_pollos}
\item \link{lluvia_forraje}
\item \link{lotes_BA}
\item \link{maiz}
\item \link{molinos}
\item \link{monedas}
\item \link{MS_girasol}
\item \link{N_suelo}
\item \link{neumaticos}
\item \link{neumaticos_tractores}
\item \link{ovejas}
\item \link{penetrometro}
\item \link{pesada_novillos}
\item \link{pesada_terneros}
\item \link{peso_balanceado}
\item \link{pezoneras}
\item \link{plantacion}
\item \link{plantacion_dap}
\item \link{P_maiz}
\item \link{pollos}
\item \link{productores}
\item \link{proteina_caseina}
\item \link{pulsaciones_caballos}
\item \link{raciones}
\item \link{razas}
\item \link{riesgo_arbolado}
\item \link{ryegrass}
\item \link{semilla}
\item \link{semillas}
\item \link{semilla_trigo}
\item \link{sorgo}
\item \link{superficies_EAP}
\item \link{susceptibilidad}
\item \link{tambos_2008_2015}
\item \link{TRC}
\item \link{tribolium}
}
}

