% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_groups.R
\docType{data}
\name{datasets_estex}
\alias{datasets_estex}
\title{Datasets para Estadística Experimental}
\description{
Lista de datasets utilizados en la asignatura Estadística Experimental.
}
\section{Datasets}{

\itemize{
\item \link{ace_codornices}
\item \link{arroz_pd}
\item \link{caballos}
\item \link{cana_azucar}
\item \link{cana_pd}
\item \link{conc_glucosa}
\item \link{conservante_frutilla}
\item \link{dieta_cerdos}
\item \link{dieta_pavos}
\item \link{dietas_novillos}
\item \link{digest_in_vitro}
\item \link{dms_ensilaje}
\item \link{energia_abejas}
\item \link{estrogeno}
\item \link{fert_melon}
\item \link{fertilizante_pastizal}
\item \link{fibra_quimico}
\item \link{fijacion_N}
\item \link{gallinas}
\item \link{gladiolos}
\item \link{guayule}
\item \link{hibridos_maiz}
\item \link{hormona_ovario}
\item \link{inhibidor_N}
\item \link{insecticida_poroto}
\item \link{invernadero_spartina}
\item \link{inyeccion_vaquillas}
\item \link{irrigacion_naranjo}
\item \link{labores_remolacha}
\item \link{larvas_cesped}
\item \link{lechuga}
\item \link{maiz2_pd}
\item \link{met_pesados}
\item \link{nabos}
\item \link{nectar_trebol_rojo}
\item \link{nematodes_durazno}
\item \link{pH_laboratorio}
\item \link{peso_vaquillonas}
\item \link{prolactina_vacas}
\item \link{raciones_pollos}
\item \link{resist_algodon}
\item \link{rollos_tela}
\item \link{sal_cebada}
\item \link{sal_suelo}
\item \link{semillas_pd}
\item \link{semillas_trigo}
\item \link{soja_pd}
\item \link{sorgo_urea}
\item \link{tiempo_reaccion}
\item \link{tomate}
\item \link{trebol_rizhobium}
\item \link{trigo_fungicidas_momentos}
\item \link{trigo_zona_ciclo}
\item \link{variedad_trigo}
\item \link{vitA_vaquillas}
}
}

