% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{dieta_cerdos}
\alias{dieta_cerdos}
\title{Efecto de la dieta en cerdos}
\format{
A data frame with 26 rows and 2 columns:
\itemize{
\item \code{D}: tipo de dieta suministrada.
\item \code{AP}: aumento de peso en kg (AP).
}
}
\source{
--
}
\usage{
data(dieta_cerdos)
}
\description{
Ensayo realizado para evaluar el efecto de 6 dietas en el aumento de peso en
kg (AP) de cerdos durante un tiempo determinado.
}
\details{
Se seleccionaron 30 cerdos de similar peso y edad y cada dieta se suministró
individualmente al azar a 5 cerdos.

Los ingredientes de las dietas evaluadas fueron:
\itemize{
\item 
\enumerate{
\item Maíz.
}
\item 
\enumerate{
\item Maíz y trigo.
}
\item 
\enumerate{
\item Maíz y cebada.
}
\item 
\enumerate{
\item Sorgo.
}
\item 
\enumerate{
\item Sorgo y trigo.
}
\item 
\enumerate{
\item Sorgo y cebada.
No obstante, durante el experimento algunas de las repeticiones tuvieron que
ser descartadas quedaron 26 datos en total.
}
}
}
\keyword{internal}
