% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{dietas_pollos}
\alias{dietas_pollos}
\title{Ganancia de peso en pollos por ración de alimento}
\format{
A data frame with 16 rows and 3 columns:
\itemize{
\item \code{pollo}: identificador numérico de la pareja de pollos
\item \code{dieta}: tipo de ración de alimento suministrada ("Tradicional" o "Especial")
\item \code{ganancia}: ganancia de peso registrada (en kg)
}
}
\usage{
data(dietas_pollos)
}
\description{
Datos de ganancia de peso (en kg) de 8 pares de pollos idénticos alimentados con ración tradicional o ración especial.
}
\details{
Se buscaron 8 pares de pollos idénticos en cuanto a peso, raza y sexo. A un lote se le suministró por 15 días el alimento tradicional y al otro lote una ración especial. Se detalla la ganancia de peso de cada animal al finalizar el periodo.
}
\keyword{internal}
