% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{digest_in_vitro}
\alias{digest_in_vitro}
\title{Digestibilidad in vitro}
\format{
A data frame with 20 rows and 4 columns:
\itemize{
\item \code{Dosis}: dosis de aditivo en g/100 gMS.
\item \code{Tubo}: id del tubo de ensayo.
\item \code{dig48}: digestibilidad a las 48 hs en \%.
\item \code{Dosis_ord}: dosis transformada a factor ordenado.
}
}
\source{
--
}
\usage{
data(digest_in_vitro)
}
\description{
Se investigó el efecto de la adición de un aditivo para aumentar la
digestibilidad de la fibra de un forraje.
}
\details{
Se prepararon tubos de ensayo con la misma cantidad de forraje y licor
ruminal, mientras que las dosis de aditivo (0, 5, 10, 15 g/100 gMS) se
asignaron al azar a cada tubo por quintuplicado.

Se midió la digestibilidad a las 48 hs.
}
\keyword{internal}
