% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{dosis_tomate}
\alias{dosis_tomate}
\title{Efecto de dosis de fertilizante en crecimiento de tomate}
\format{
A data frame with 11 rows and 3 columns:
\itemize{
\item \code{dosis}: dosis de fertilizante aplicada (ml/ha)
\item \code{fruto}: rendimiento de fruto obtenido (en miles de kg/ha de peso húmedo)
\item \code{tallo}: número de tallos por metro cuadrado (m2)
}
}
\usage{
data(dosis_tomate)
}
\description{
Datos de rendimiento de fruto y cantidad de tallos por metro cuadrado registrados en plantas de tomate bajo distintas dosis de fertilizante.
}
\details{
Se realizó un análisis de crecimiento de tomate en invernadero y se midieron distintas variables como rendimiento de fruto en kg/ha y la cantidad de tallos por m2. Las plantas fueron establecidas en macetas, se utilizó un sistema de riego dirigido y se les aplicaron diferentes dosis de fertilizante. Al momento de la cosecha se contabilizaron los tallos y se pesaron los frutos.
}
\keyword{internal}
