% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{especies}
\alias{especies}
\title{Especies}
\format{
A data frame with 6 rows and 3 columns:
\itemize{
\item \code{Especie}: Especie
\item \code{Hojas}: Estado de las hojas (sanas, enfermas)
\item \code{Número}: Número de ejemplares
}
}
\usage{
data(especies)
}
\description{
Datos de cantidad de hojas sanas y enfermas según especie.
}
\keyword{internal}
