% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{establecimientos}
\alias{establecimientos}
\title{Superficie de establecimientos agropecuarios}
\format{
A data frame with 12 rows and 1 column:
\itemize{
\item \code{superficie}: superficie del establecimiento en hectáreas (ha)
}
}
\usage{
data(establecimientos)
}
\description{
Datos de superficie (en hectáreas) de una muestra de 12 establecimientos agropecuarios de una región.
}
\details{
Un estudio económico desea probar si la superficie media de los establecimientos de una región es mayor a 135 ha. Para ello se tomó una muestra de 12 establecimientos agropecuarios y se obtuvieron los siguientes datos de superficie en ha.
}
\keyword{internal}
