% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{estrogeno}
\alias{estrogeno}
\title{Actividad estrogénica}
\format{
A data frame with 28 rows and 2 columns:
\itemize{
\item \code{SO}: solucion de estrógeno asignada.
\item \code{PE}: Peso del útero.
}
}
\source{
--
}
\usage{
data(estrogeno)
}
\description{
Experimento realizado para evaluar la actividad estrogénica de 6 tipos de
solución de estrógenos sujetas a una técnica de inactivación in vitro.
}
\details{
Las 6 soluciones y un tratamiento control se asignaron por cuadriplicado a
ratones hembra.
Se registró el peso (PE) del útero como medida de la actividad estrogénica.
}
\keyword{internal}
