% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{fert_melon}
\alias{fert_melon}
\title{Fertilización en melón}
\format{
A data frame with 16 rows and 4 columns:
\itemize{
\item \code{Fila}: id de la fila del diseño.
\item \code{Column}: id de la columna del diseño.
\item \code{Tratamiento}: tratamiento aplicado.
\item \code{rend}: peso seco de las plantas de melón.
}
}
\source{
--
}
\usage{
data(fert_melon)
}
\description{
Experimento para comparar dos formas de ubicar el fertilizante (en banda y al
voleo) y dos mezclas de fertilizante (A y B) en los rendimientos de melón.
}
\details{
Los tratamientos fueron aplicados a parcelas pequeñas en un diseño de
Cuadrado Latino (filas y columnas) debido a la variabilidad del suelo y
pendiente.

Tratamientos:
\itemize{
\item T1: al voleo fert A
\item T2: al voleo fert B
\item T3: en banda fert A
\item T4: en banda fert B
}
}
\keyword{internal}
