% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{gallinas}
\alias{gallinas}
\title{Función pituitaria en gallinas}
\format{
A data frame with 25 rows and 2 columns:
\itemize{
\item \code{TR}: regimen de muda suministrado.
\item \code{T3}: concentración de suero T3.
}
}
\source{
--
}
\usage{
data(gallinas)
}
\description{
Un fisiólogo de animales estudió la función pituitaria de las gallinas
sometidas a un régimen estándar de muda de pluma forzada usado por los
productores de huevo para mantener las gallinas en producción.
}
\details{
Se usaron 25 gallinas en el estudio. Cinco se utilizaron para la medición
previa al régimen de muda forzada y el resto al final de cada una de las
cuatro etapas del régimen.

Las cinco etapas del régimen (TR) fueron:
\itemize{
\item 
\enumerate{
\item premuda (control).
}
\item 
\enumerate{
\item ayuno de 8 días.
}
\item 
\enumerate{
\item 60 g de salvado al día durante 10 días.
}
\item 
\enumerate{
\item 80 g de salvado al día por 10 días.
}
\item 
\enumerate{
\item mezcla de malta durante 42 días.
Uno de los compuestos medidos fue la concentración de suero T3. Los datos de
la tabla son las medidas de suero T3 en las cinco gallinas sacrificadas al
final de cada etapa del régimen.
}
}
}
\keyword{internal}
