% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{ganancia_novillos}
\alias{ganancia_novillos}
\title{Ganancia de peso en novillos Holando con suplementación mineral}
\format{
A data frame with 28 rows and 3 columns:
\itemize{
\item \code{novillo}: identificador numérico del animal
\item \code{tratamiento}: tipo de tratamiento recibido ("Placebo" o "Mineral")
\item \code{ganancia}: ganancia de peso registrada (en kg)
}
}
\usage{
data(ganancia_novillos)
}
\description{
Datos de ganancia de peso (en kg) de novillos de raza Holando sometidos a un tratamiento con suplementación mineral o placebo.
}
\details{
Se realizó un ensayo con novillos de raza Holando. A un grupo de animales se le dio un mineral y al otro grupo un placebo. Se detalla la ganancia de peso (kg) acumulada por cada animal durante el ensayo.
}
\keyword{internal}
