% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{glifosato}
\alias{glifosato}
\title{Peso de empaque de cajas de Glifosato}
\format{
A data frame with 1150 rows and 3 columns:
\itemize{
\item \code{Mes}: mes de registro del empaque (ej. "enero", "febrero", etc.)
\item \code{Envase}: identificador numérico de la caja/envase
\item \code{Peso}: peso del empaque en kilogramos (kg)
}
}
\usage{
data(glifosato)
}
\description{
Datos de peso de empaques de cajas de Glifosato (herbicida granulado soluble) registrados para diferentes meses y envases, utilizados para control de calidad.
}
\details{
Datos correspondientes a mediciones de control de calidad del peso de empaque para cajas de Glifosato (un herbicida granulado soluble que se comercializa en paquetes de 15 kg). Incluye registros de múltiples meses y envases con el objetivo de evaluar el funcionamiento de la maquinaria de empaque y controlar la variabilidad y desviaciones en los pesos de carga de los envases.
}
\keyword{internal}
