% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{hormona_ovario}
\alias{hormona_ovario}
\title{Hormona foliculoestimulante en vacas}
\format{
A data frame with 16 rows and 4 columns:
\itemize{
\item \code{F}: id de la camada (filas).
\item \code{C}: categoría de peso corporal (columnas).
\item \code{HORMONA}: hormona inyectada.
\item \code{PESO}: peso del ovario medido.
}
}
\source{
--
}
\usage{
data(hormona_ovario)
}
\description{
La actividad de la hormona foliculoestimulante en vacas fue medida mediante
la respuesta del peso de los ovarios de ratas inmaduras.
}
\details{
Los factores conocidos que afectan la respuesta son el genotipo y el peso
corporal. Asumiendo que la relación entre el peso de los ovarios y el peso
corporal es independiente de las diferencias genéticas, se utilizó un diseño
Cuadrado Latino donde las filas son las camadas de ratas y las columnas las
categorías de peso corporal.

Las hormonas fueron inyectadas a las ratas y luego de un período de tiempo se
midió el peso del ovario.
}
\keyword{internal}
