% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{insectos}
\alias{insectos}
\title{Distribución de insectos en plantas}
\format{
A data frame with 6 rows and 2 columns:
\itemize{
\item \code{obs}: número de plantas observadas con la cantidad de insectos correspondiente (frecuencia)
\item \code{insectos}: número de insectos por planta (0 a 5)
}
}
\usage{
data(insectos)
}
\description{
Distribución de frecuencia del número de insectos observados por planta en una muestra de 400 plantas.
}
\details{
Datos correspondientes a la distribución del número de insectos en una muestra de 400 plantas de cultivo.
}
\keyword{internal}
