% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.local_repo.R
\name{install.local_repo}
\alias{install.local_repo}
\title{Instalar desde miniCRAN}
\usage{
install.local_repo(pkgs = NULL, repo_path = NULL, install = T)
}
\arguments{
\item{pkgs}{Espera un vector de caracteres con los paquetes a instalar}

\item{repo_path}{Espera un string con la ruta donde instalar el repositorio.
Si no se especifica se toma el del script donde se está ejecutando}

\item{install}{Parámetro de debug que permite correr la función sin
descargar nada}
}
\value{
Retorna una lista con la ruta del repositorio, el tipo de paquetes
descargados, los paquetes exitosos y los que dieron error
}
\description{
Función que instala los paquetes especificados desde un
repositorio local
}
\details{
Primero chequea que se halla ingresado una lista de paquetes y el SO
en el que opera
Luego chequea si se especifico la ruta donde deberá descargarse el
repositorio, si no lo está utiliza la del scrip donde es está ejecutando
Seguido de esto chequea cuales de los paquetes indicados están dentro del
repositorio
Por ultimo instala los paquetes validos y retorna un informe con la ruta del
repositorio, los paquetes validos y los invalidos
}
\examples{
pkgs <- c(
    "rmarkdown", "moments", "pacman", "rio", "tidyverse",
    "emmeans", "car", "DescTools", "agricolae", "multcomp", "multcompView",
    "ggplot2"
)
\dontrun{
install.local_repo(pkgs)
}
}
