% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{inyeccion_vaquillas}
\alias{inyeccion_vaquillas}
\title{Métodos de inyección en vaquillas}
\format{
A data frame with 16 rows and 6 columns:
\itemize{
\item \code{F}: línea genética (filas).
\item \code{C}: método de inyección (columnas).
\item \code{D}: dosis aplicada (A, B, C, D).
\item \code{RTA}: aumento de peso en kg.
\item \code{D_num}: dosis aplicada en formato numerico.
\item \code{D_ord}: dosis aplicada como factor ordenado.
}
}
\source{
--
}
\usage{
data(inyeccion_vaquillas)
}
\description{
Se evaluó la respuesta a cuatro dosis (D) diferentes de un producto
inyectado en vaquillas Holstein para lograr el incremento en su peso.
}
\details{
Para controlar el efecto de la genética y el tipo de aplicación, se utilizó
un diseño cuadrado latino donde por fila se agruparon las vaquillas según su
línea genética y las columnas representan el método de inyección (1 =
intravenosa, 2 = intramuscular, 3 = subcutánea I y 4 = subcutánea II).

La respuesta se midió en aumento de peso (kg) en un determinado periodo de
tiempo.

Las dosis fueron:
\itemize{
\item A: 0 ml
\item B: 2 ml
\item C: 4 ml
\item D: 6 ml
}
}
\keyword{internal}
