% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{irrigacion_naranjo}
\alias{irrigacion_naranjo}
\title{Irrigación en naranjos}
\format{
A data frame with 48 rows and 2 columns:
\itemize{
\item \code{BL}: id del bloque.
\item \code{TR}: tipo de riego utilizado.
\item \code{PF}: peso de la fruta cosechada en libras.
}
}
\source{
--
}
\usage{
data(irrigacion_naranjo)
}
\description{
Efecto de diferentes tipos de irrigación en el peso de la fruta cosechada en
naranjos.
}
\details{
En un naranjal se llevo a cabo un experimento de irrigación con un diseño de
bloques completos aleatorizado en el que se usaron seis tratamientos de
irrigación:
\itemize{
\item G = goteo.
\item E = por surco.
\item R = rocío.
\item A = aspersor.
\item AR = aspersor + rocío.
\item AN = anegado.
}

Los siguientes datos se refieren al peso en libras de la fruta cosechada en cada parcela.
}
\keyword{internal}
