% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{labores_remolacha}
\alias{labores_remolacha}
\title{Labores culturales en remolacha azucarera}
\format{
A data frame with 9 rows and 4 columns:
\itemize{
\item \code{F}: id de la fila.
\item \code{C}: id de la columna.
\item \code{LAB}: tipo de labor cultural aplicada.
\item \code{RTO}: rendimiento en toneladas por hectárea.
}
}
\source{
--
}
\usage{
data(labores_remolacha)
}
\description{
Se midieron los rendimientos de remolacha azucarera en toneladas por
hectárea bajo tres tipos de labores culturales en un diseño en cuadrado
latino.
}
\details{
--
}
\keyword{internal}
