% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{lechuga}
\alias{lechuga}
\title{Respuesta de la lechuga a la fertilización nitrogenada}
\format{
A data frame with 20 rows and 24 columns:
\itemize{
\item \code{Dosis}: dosis de nitrogeno aplicado.
\item \code{Dosis_ord}: dosis transformada a factor ordenada.
\item \code{Dosis_chr}: dosis como caracter transformada a factor.
\item \code{Plantas}: numero de plantas cosechadas de la parcela.
}
}
\source{
--
}
\usage{
data(lechuga)
}
\description{
Experimento realizado para evaluar la respuesta en el stand de plantas de
lechuga a la fertilización nitrogenada.
}
\details{
Se aplicaron cinco dosis de nitrato de amonio (0, 50, 100, 150 y 200 kg/ha)
por cuadruplicado a parcelas siguiendo un diseño completamente aleatorizado.

Se registró el número de lechugas cosechadas de la parcela.
}
\keyword{internal}
