% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{maiz}
\alias{maiz}
\title{Densidad de plantas en lote de maíz}
\format{
A data frame with 6 rows and 2 columns:
\itemize{
\item \code{plantas}: número de plantas por metro cuadrado
\item \code{parcelas}: cantidad de sectores (frecuencia)
}
}
\usage{
data(maiz)
}
\description{
Datos de frecuencia del número de plantas por metro cuadrado nacidas en sectores dentro de un lote de maíz.
}
\details{
Las observaciones corresponden al número de plantas \eqn{m^{-2}} nacidas en 20 sectores dentro de un lote de maíz.
}
\keyword{internal}
