% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{maiz2_pd}
\alias{maiz2_pd}
\title{Abonos en maíz}
\format{
A data frame with 36 rows and 4 columns:
\itemize{
\item \code{Bloque}: id del bloque.
\item \code{Aplicacion}: forma de aplicación (parcela principal).
\item \code{Abono}: tipo de abono (subparcela).
\item \code{Rend}: rendimiento por parcela en kg.
}
}
\source{
--
}
\usage{
data(maiz2_pd)
}
\description{
Se probó el efecto de seis abonos y dos formas de Aplicación, superficial (1)
o Incorporado (2) en maíz.
}
\details{
Diseño en parcelas divididas (DPD) en bloques completamente al azar en donde
las variantes de aplicación se asignaron a las parcelas (principal) y los
abonos a las subparcelas.
}
\keyword{internal}
