% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkminiCRAN.R
\name{mkminiCRAN}
\alias{mkminiCRAN}
\title{Armar miniCRAN local}
\usage{
mkminiCRAN(
  pkgs,
  path = NULL,
  so = c("windows", "linux", "macos"),
  mirrors = NULL
)
}
\arguments{
\item{pkgs}{Vector de caracteres. Espera los paquetes a instalar}

\item{path}{String. Espera la ruta donde deberá armarse el miniCRAN}

\item{so}{Vector de caracteres. Espera los sistemas operativos a
utilizar. Por defecto incluye los tres más importantes}

\item{mirrors}{Vector de caracteres. Puede especificarse una lista de mirrors a utilizar. Si no se
especifica usa los que trae por defecto}
}
\value{
Devuelve una vector con los paquetes instalados+
}
\description{
Función que arma un miniCRAN de manera local
}
\details{
Primero chequea el mirror a utilizar.
Luego crea el directorio donde almacenar el repositorio.
Luego obtiene las dependencias de los paquetes de acuerdo al SO a utilizar
Por ultimo descarga los paquetes necesarios y retorna un vector con los
paquetes descargados.
}
\examples{
pkgs <- c(
    "rmarkdown", "moments", "pacman", "rio", "tidyverse", "emmeans",
    "car", "DescTools", "agricolae", "multcomp", "multcompView", "ggplot2"
)
path <- "/home/path_to_repo"
\dontrun{
mkminiCRAN(pkgs, path)
}
}
