% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{monedas}
\alias{monedas}
\title{Lanzamiento de cuatro monedas}
\format{
A data frame with 5 rows and 2 columns:
\itemize{
\item \code{obs}: frecuencia de lanzamientos observados con la cantidad de caras dada
\item \code{caras}: cantidad de caras obtenidas en el lanzamiento (0 a 4)
}
}
\usage{
data(monedas)
}
\description{
Frecuencias observadas del número de caras obtenidas en el lanzamiento simultáneo de 4 monedas repetido 160 veces.
}
\details{
Distribución de resultados obtenidos al lanzar un grupo de 4 monedas de forma simultánea un total de 160 veces, registrando el número de caras.
}
\keyword{internal}
