% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{N_suelo}
\alias{N_suelo}
\title{Contenido de nitrógeno nítrico en suelos}
\format{
A data frame with 50 rows and 3 columns:
\itemize{
\item \code{suelo}: identificador numérico de la muestra de suelo
\item \code{tipo}: tipo o calidad del suelo ("Bueno" o "Pobre")
\item \code{ppm}: contenido de nitrógeno nítrico (en partes por millón, ppm)
}
}
\usage{
data(N_suelo)
}
\description{
Datos de contenido de nitrógeno nítrico (N-NO3) en partes por millón (ppm) para dos tipos de suelos (Bueno y Pobre) del norte de la provincia de Santa Fe.
}
\details{
El contenido de N-NO3 (ppm) en diferentes suelos del norte de la provincia de Santa Fe fue medido y clasificado según el tipo de suelo (Bueno y Pobre).
}
\keyword{internal}
