% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{nabos}
\alias{nabos}
\title{Ácido ascórbico en nabos}
\format{
A data frame with 15 rows and 3 columns:
\itemize{
\item \code{BL}: id del bloque.
\item \code{TR}: tipo de lavado y secado aplicado.
\item \code{LE}: contenido de ácido ascórbico.
}
}
\source{
--
}
\usage{
data(nabos)
}
\description{
Estudio realizado para determinar el efecto de lavar y eliminar el exceso de
humedad secando o mediante corriente de aire sobre el contenido de ácido
ascórbico de nabos.
}
\details{
Se presentan los datos de los 3 tratamientos
\itemize{
\item C = control sin lavar.
\item LC = lavado con agua corriente.
\item LSA = lavado y secado con absorbente.
\item LSC = lavado y secado con corriente de aire.
Los tratamientos fueron asignados al azar dentro de 5 bloques
}
}
\keyword{internal}
