% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{nematodes_durazno}
\alias{nematodes_durazno}
\title{Nematodes en durazno}
\format{
A data frame with 18 rows and 2 columns:
\itemize{
\item \code{Tratamiento}: tipo de tratamiento aplicado.
\item \code{altura}: altura del plantin en cm.
}
}
\source{
--
}
\usage{
data(nematodes_durazno)
}
\description{
Experimento realizado para comparar el crecimiento de plantines de durazno.
Investigadores suponen que la enfermedad tristeza de los tallos en plantines
de árboles de durazno puede estar relacionada con la presencia o ausencia de
nemátodes en el suelo.

El control de malezas y el tratamiento de suelos usando herbicidas puede ser
efectivo en promover el crecimiento de los plantines.
}
\details{
Se trato el suelo y las malezas con uno de los siguientes tratamientos:
\itemize{
\item A: Control (sin herbicida).
\item B: Herbicida con Nemagone (nematicida).
\item C: Herbicida sin Nemagone (nematicida).
Los investigadores asignaron al azar 6 de los 18 plantines a cada
tratamiento, luego trataron el suelo y las malezas en las áreas de
crecimiento con el tratamiento asignado.
}

Al final del estudio, registraron la altura (en cm) de cada plantin.
}
\keyword{internal}
