% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{neumaticos_tractores}
\alias{neumaticos_tractores}
\title{Duración de neumáticos de tractores}
\format{
A data frame with 15 rows and 1 column:
\itemize{
\item \code{duracion}: duración del neumático en miles de kilómetros (mil km)
}
}
\usage{
data(neumaticos_tractores)
}
\description{
Datos de duración (en miles de kilómetros) de una muestra de 15 neumáticos de tractores sometidos a un ensayo.
}
\details{
Un fabricante de neumáticos de tractores mide en miles de km la duración de los mismos. En un ensayo de 15 neumáticos la duración de los mismos fue registrada en este set de datos.
}
\keyword{internal}
