% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{P_maiz}
\alias{P_maiz}
\title{Relación de fósforo inicial en suelo y recuperado por plantas de maíz}
\format{
A data frame with 9 rows and 2 columns:
\itemize{
\item \code{P_suelo}: cantidad de fósforo inorgánico inicial en el suelo (ppm)
\item \code{P_planta}: cantidad de fósforo recuperado en la planta a los 30 días (ppm)
}
}
\usage{
data(P_maiz)
}
\description{
Datos de concentraciones de fósforo (P) inorgánico en suelo al inicio del cultivo y la cantidad de fósforo recuperado por plantas de maíz a los 30 días.
}
\details{
Se realizó un experimento para determinar la relación entre la cantidad de P inorgánico en el suelo (ppm) al inicio del cultivo de maíz y la cantidad de P que es recuperada por la planta (en ppm) a los 30 días. Se prepararon macetas con distintas concentraciones de P inicial, se sembró el mismo híbrido de maíz y luego de 30 días se determinó el contenido de P en la planta.
}
\keyword{internal}
