% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{penetrometro}
\alias{penetrometro}
\title{Resistencia mecánica a la penetración de suelo}
\format{
A data frame with 2000 rows and 3 columns:
\itemize{
\item \code{Muestra}: identificador numérico del punto de muestreo
\item \code{RP}: resistencia mecánica a la penetración (en megapascales, MPa)
\item \code{momento}: momento del muestreo en relación a la labranza ("antes", "despues")
}
}
\usage{
data(penetrometro)
}
\description{
Datos de resistencia mecánica a la penetración (RP) en megapascales (MPa) medidos en muestras de suelo en diferentes momentos relativos a una labor de labranza.
}
\details{
Datos correspondientes a mediciones de resistencia mecánica a la penetración (RP) de un suelo, expresadas en megapascales (MPa). La resistencia mecánica es una propiedad del suelo crítica para el desarrollo radical, existiendo un umbral de 2 MPa que puede limitar el cultivo y requerir labranza. El set de datos incluye determinaciones de RP en diferentes muestras (ubicaciones) tomadas antes y después de realizar una labor de labranza vertical.
}
\keyword{internal}
