% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{pesada_novillos}
\alias{pesada_novillos}
\title{Datos del pesaje de dos tropas de novillos extraídos del sistema de registros del establecimiento "La Escondida S.A."}
\format{
A data frame with 1842 rows and 13 columns:
\itemize{
\item \code{IDV}: código de identificación visual (caravana)
\item \code{Tropa}: número de tropa al que pertenece.
\item \code{Procedencia}: nombre del establecimiento del cual proviene.
\item \code{Fecha_Ingreso}: fecha de ingreso al sistema de engorde.
\item \code{Peso_inicial}: peso al ingresar al sistema de engorde.
\item \code{Peso_anterior}: peso registrado en la pesada anterior.
\item \code{GPV_anterior}: ganancia de peso vivo en la ultima pesada, i.e. diferencia de peso entre las últimas dos pesadas.
\item \code{GDM_anterior}: ganancia de peso diaria promedio anterior
\item \code{Fecha}: fecha de la pesada actual (MM/DD/YYYY)
\item \code{Hora}: hora en la que se realizó la pesada actual (HH:MM:SS)
\item \code{Peso}: peso actual registrado en kg.
\item \code{Días}: número de días entre la pesada anterior y la actual.
\item \code{Días_total}: número de dias desde el ingreso al sistema.
}
}
\source{
Gentileza Ing. Agr. Marcón
}
\usage{
data(pesada_novillos)
}
\description{
Datos del pesaje de dos tropas de novillos extraídos del sistema de registros del establecimiento "La Escondida S.A."
}
\keyword{internal}
