% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{pesada_terneros}
\alias{pesada_terneros}
\title{Pesaje de 1598 terneros que ingresaron al sistema de engorde de la estancia "La Escondida S.A."}
\format{
A data frame with 1598 rows and 8 columns:
\itemize{
\item \code{Procedencia}: nombre del establecimiento del cual provienen los terneros de destete.
\item \code{IDV}: código de identificación visual (caravana)
\item \code{Sexo}: sexo del animal
\item \verb{Fecha Ingreso}: fecha de ingreso al sistema de engorde
\item \code{Categoria}: categoría dentro a la que pertenecen dentro del rodeo de engorde.
\item \code{Fecha}: fecha en la que se realizó la pesada (MM/DD/YYYY)
\item \code{Hora}: hora en la que se realizó la pesada (HH:MM:SS)
\item \code{Peso}: peso del animal en kg.
}
}
\source{
Gentileza Ing. Agr. Marcón
}
\usage{
data(pesada_terneros)
}
\description{
Pesaje de 1598 terneros que ingresaron al sistema de engorde de la estancia "La Escondida S.A."
}
\keyword{internal}
