% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{peso_balanceado}
\alias{peso_balanceado}
\title{Pesos de bolsas de alimento balanceado para cerdos}
\format{
A data frame with 55 rows and 3 columns:
\itemize{
\item \code{muestra}: identificador numérico de la bolsa (muestra)
\item \code{dia}: día en que se tomó la muestra ("muestra1" o "muestra2")
\item \code{peso}: peso neto de la bolsa de alimento balanceado (en kg)
}
}
\usage{
data(peso_balanceado)
}
\description{
Datos de pesos netos (en kg) de bolsas de alimento balanceado para cerdos tomadas en dos días de producción.
}
\details{
Se sabe que una máquina de empacar balanceado para cerdos vierte el balanceado en bolsas de 20 kg, con una desviación estándar de 3.8 kg. Se llevan a cabo verificaciones constantes de los pesos netos de las bolsas para mantener el ajuste de la maquinaria que controla el peso. Dos muestras tomadas en dos días presentan la información registrada en este set de datos.
}
\keyword{internal}
