% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{peso_vaquillonas}
\alias{peso_vaquillonas}
\title{Anabólicos en vaquillonas}
\format{
A data frame with 28 rows and 2 columns:
\itemize{
\item \code{AN}: tipo de anabólico suministrado.
\item \code{AP}: aumento de peso en kg.
}
}
\source{
--
}
\usage{
data(peso_vaquillonas)
}
\description{
Estudió del efecto de anabólicos (AN) en el aumento de peso (AP) en kg de
vaquillonas.
}
\details{
Se utilizaron en total 12 vaquillonas de similares características (peso,
edad, etc).

Cuatro vaquillonas elegidas al azar recibieron un tipo de anabólico mientras
que otras 4 vaquillonas recibieron el otro tipo de anabólico. Adicionalmente,
4 vaquillonas fueron dejadas como control sin recibir anabólicos.

Durante el experimento, una de las vaquillonas del tratamiento 2 tuvo que ser
descartada del ensayo.
}
\keyword{internal}
