% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{pezoneras}
\alias{pezoneras}
\title{Duración de pezoneras de ordeño}
\format{
A data frame with 20 rows and 1 column:
\itemize{
\item \code{duracion}: duración de la pezonera en horas
}
}
\usage{
data(pezoneras)
}
\description{
Datos de duración (en horas) de una muestra de 20 pezoneras de ordeño, utilizados para evaluar la afirmación del fabricante sobre su vida útil media.
}
\details{
Un fabricante de pezoneras sostiene que la duración media de las mismas es superior a 1400 h. Se tomó una muestra de 20 pezoneras y se detalla su duración en este set de datos.
}
\keyword{internal}
