% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{pH_laboratorio}
\alias{pH_laboratorio}
\title{Error en pH-metros}
\format{
A data frame with 24 rows and 3 columns:
\itemize{
\item \code{Dispositivo}: letra que identifica el dispositivo utilizado.
\item \code{Muestra}: id de la muestra a la que se le realizo la medida.
\item \code{dif}: diferencia de pH medido vs pH conocido.
}
}
\source{
--
}
\usage{
data(pH_laboratorio)
}
\description{
Un laboratorio de suelos tiene 4 tipos de pH-metros para determinar el pH en
muestras de suelo. El encargado del laboratorio quiere determinar si hay
diferencias en el promedio de las lecturas obtenidas por estos dispositivos.
}
\details{
Para ello selecciona 24 muestras de suelo con pH conocido y asigna
aleatoriamente seis muestras a cada pH-metro. Cada muestra de suelo es
analizada y se registra la diferencia entre el valor que lee el pH-metro y
el valor conocido de pH de la muestra.
}
\keyword{internal}
