% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{plantacion}
\alias{plantacion}
\title{Altura de ejemplares de Eucalyptus}
\format{
A data frame with 10 rows and 1 column:
\itemize{
\item \code{altura}: Altura del árbol en metros
}
}
\usage{
data(plantacion)
}
\description{
Datos de altura en metros de una muestra de 10 ejemplares de árboles \emph{Eucalyptus} en su segundo año de plantación de un predio de San Agustín (Santa Fe).
}
\keyword{internal}
