% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{prolactina_vacas}
\alias{prolactina_vacas}
\title{Prolactina en vaquillas}
\format{
A data frame with 9 rows and 6 columns:
\itemize{
\item \code{F}: factor de bloqueo peso (filas).
\item \code{C}: factor de bloqueo genética (columnas).
\item \code{DOSIS}: dosis de hormona TRH aplicada.
\item \code{PICO}: pico máximo de prolactina medido.
\item \code{DOSIS_num}: dosis aplicada en formato numerico.
\item \code{DOSIS_ord}: dosis aplicada como factor ordenado.
}
}
\source{
--
}
\usage{
data(prolactina_vacas)
}
\description{
Se realizó un experimento para medir el pico maximo de prolactina en
vaquillas Holstein ante la aplicacion de distintas dosis de la hormona TRH.
}
\details{
Se utilizaron 3 dosis de la hormona: A = 50 μg, B = 100 μg y C = 150 μg en
un DCL. Los factores de bloqueo fueron genética (columnas) y el peso (filas).
}
\keyword{internal}
