% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{pulsaciones_caballos}
\alias{pulsaciones_caballos}
\title{Pulsaciones del corazón de caballos de carrera}
\format{
A data frame with 15 rows and 1 column:
\itemize{
\item \code{pulsaciones}: Pulsaciones del corazón por minuto (latidos por minuto)
}
}
\usage{
data(pulsaciones_caballos)
}
\description{
Datos de pulsaciones del corazón por minuto de una muestra de 15 caballos de carrera registrados en un estudio.
}
\keyword{internal}
