% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{raciones}
\alias{raciones}
\title{Ganancia diaria de peso en novillos por tipo de ración}
\format{
A data frame with 24 rows and 3 columns:
\itemize{
\item \code{novillo}: identificador numérico del animal
\item \code{racion}: ración alimenticia suministrada ("1" o "2", correspondientes a A y B)
\item \code{peso}: ganancia diaria de peso por animal (en kg/día)
}
}
\usage{
data(raciones)
}
\description{
Datos de ganancia de peso diaria promedio por animal (en kg) en novillos sometidos a dos raciones alimenticias diferentes (A y B).
}
\details{
En un ensayo de engorde de novillos se utilizaron dos raciones (A y B) en dos lotes de 12 animales cada uno. La variable respuesta fue la ganancia de peso por animal por día (kg). Por información previa se consideró que las varianzas poblacionales eran iguales, con un valor de 0.65.
}
\keyword{internal}
