% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{razas}
\alias{razas}
\title{Relación entre tipo de parto y raza en vacas lecheras}
\format{
A data frame with 6 rows and 3 columns:
\itemize{
\item \code{Tipo}: tipo de parto ("normal" o "anormal")
\item \code{Raza}: raza de la vaca ("Holstein", "Cruza", "Jersey")
\item \code{partos}: número de partos registrados (frecuencia)
}
}
\usage{
data(razas)
}
\description{
Frecuencias observadas de tipo de parto (normal o anormal) según la raza en una muestra de 300 vacas de una cuenca lechera.
}
\details{
Para estudiar la relación entre el tipo de parto y la raza en vacas lecheras se tomó una muestra de 300 vacas de una cuenca lechera y se las clasificó según dichos atributos.
}
\keyword{internal}
