% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{resist_algodon}
\alias{resist_algodon}
\title{Resistencia de fibra sintética con algodón}
\format{
A data frame with 25 rows and 3 columns:
\itemize{
\item \code{BL}: id del bloque.
\item \code{AL}: porcentaje de algodón en la fibra.
\item \code{AL_ord}: porcentaje de algodón transformado a factor ordenado.
\item \code{RT}: resistencia a la tensión registrada.
}
}
\source{
--
}
\usage{
data(resist_algodon)
}
\description{
Experimento en el cual se evalúa la resistencia a la tensión de una fibra
sintética que es de interés del fabricante de la misma. Se sospecha que la
resistencia se relaciona con el porcentaje de algodón en la fibra.
}
\details{
Para ello se emplean cinco niveles de porcentaje de algodón (AL), y se
ejecutan cinco réplicas en orden aleatorio (bloques).
}
\keyword{internal}
