% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{rollos_tela}
\alias{rollos_tela}
\title{Resistencia a la tensión de rollos de tela}
\format{
A data frame with 20 rows and 3 columns:
\itemize{
\item \code{BL}: id del bloque.
\item \code{AQ}: tipo de agente químico aplicado.
\item \code{RT}: resistencia a la tensión registrada.
}
}
\source{
--
}
\usage{
data(rollos_tela)
}
\description{
Ensayo que evalua la resistencia a la tensión de rollos de algodón,
considerando los rollos de tela como bloques.
}
\details{
Se seleccionan 5 rollos y les aplica los 4 agentes químicos (AQ) de manera
aleatoria.

Se proporcionan los resultados de la resistencia a la tensión (RT).
}
\keyword{internal}
