% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{ryegrass}
\alias{ryegrass}
\title{Longitud de espigas de Ryegrass}
\format{
A data frame with 75 rows and 2 columns:
\itemize{
\item \code{planta}: número de planta
\item \code{longitud}: Longitud de la espiga en cm
}
}
\usage{
data(ryegrass)
}
\description{
En un estudio de mejoramiento se midió la longitud de 75 espigas de ryegrass.
}
\details{
La variable en estudio es la longitud de espigas (cm).
}
\keyword{internal}
