% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{sal_cebada}
\alias{sal_cebada}
\title{Salinidad en cebada}
\format{
A data frame with 18 rows and 6 columns:
\itemize{
\item \code{Salinidad}: nivel de salinidad en barios.
\item \code{Edad}: edad de la planta al momento de la cosecha en días.
\item \code{Consumo}: cantidad de agua consumida (mm/100g peso seco).
\item \code{repeticion}: id de la repetición.
\item \code{Salinidad_ord}: salinidad transformada a factor ordenado.
\item \code{Edad_ord}: edad transformada a factor ordenado.
}
}
\source{
--
}
\usage{
data(sal_cebada)
}
\description{
Experimento con plantas de cebada para medir el efecto del aumento de la
salinidad y la edad de la planta en la cantidad de agua consumida.
}
\details{
Las plantas se cultivaron en soluciones de nutrientes con 3 niveles de
salinidad (0, 6 y 12 barios) y se cosecharon a los 14, 21 y 28 días.

La medida tomada fue la cantidad de agua consumida por las plantas durante el
experimento, expresada en milímetros de agua por cada 100 gramos de peso de
la planta seca.
}
\keyword{internal}
