% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{semilla}
\alias{semilla}
\title{Variación en la forma de la semilla de plantas de Mendel}
\format{
A data frame with 10 rows and 2 columns:
\itemize{
\item \code{redonda}: cantidad de semillas con forma redonda obtenidas por planta
\item \code{angular}: cantidad de semillas con forma angular obtenidas por planta
}
}
\usage{
data(semilla)
}
\description{
Resultados de variación de planta a planta en la forma de la semilla (redonda o angular) en las primeras diez plantas de un estudio clásico de Mendel.
}
\details{
Mendel, en uno de sus clásicos estudios genéticos, observó la variación de planta a planta en una serie de experimentos sobre la forma de la semilla. Las primeras diez plantas dieron los resultados registrados en este set de datos.
}
\keyword{internal}
