% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{sorgo_urea}
\alias{sorgo_urea}
\title{Dosis de urea en sorgo granífero}
\format{
A data frame with 16 rows and 4 columns:
\itemize{
\item \code{BL}: id del bloque.
\item \code{DO}: dosis de urea aplicada.
\item \code{DO_ord}: dosis de urea transformada a factor ordenada.
\item \code{RTO}: rendimiento de grano en kg
}
}
\source{
--
}
\usage{
data(sorgo_urea)
}
\description{
En una estación experimental de la Provincia de la Pampa se realizo un
experimento con sorgo granífero.
}
\details{
Se estudiaron 3 dosis de urea (DO en kg/ha) con un testigo, obteniéndose
siguientes rendimientos de grano en Kg por parcela de 20 m^2.
}
\keyword{internal}
