% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{susceptibilidad}
\alias{susceptibilidad}
\title{Relación entre largo de semilla y susceptibilidad a roya en trigo}
\format{
A data frame with 6 rows and 3 columns:
\itemize{
\item \code{Roya}: nivel de susceptibilidad/resistencia a la roya ("resistente", "susceptible")
\item \code{largo}: longitud de la semilla ("corto", "medio", "largo")
\item \code{freq}: cantidad de semillas observadas (frecuencia)
}
}
\usage{
data(susceptibilidad)
}
\description{
Frecuencias de cruzamiento en trigo clasificadas según la longitud de la semilla y la susceptibilidad a roya.
}
\details{
Se realizó una experiencia en trigo a fin de establecer si existe relación entre los caracteres largo de semilla y susceptibilidad a roya. Se obtuvieron los valores para un cruzamiento entre un padre de semilla larga y resistente y otro de semilla corta y susceptible.
}
\keyword{internal}
