% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{tambos_2008_2015}
\alias{tambos_2008_2015}
\title{Datos sobre numero de establecimientos lecheros y composición de rodeo en la provincia de Santa Fe durante el período 2008-2015.}
\format{
A data frame with 152 rows and 10 columns:
\itemize{
\item \code{Departamento}: nombre del departamento
\item \code{Establecimientos}: número de establecimientos tamberos.
\item \verb{Vacas, Vaquillonas, Toros, Terneros, Terneras, Novillos, Novillitos}: número de cabezas según categoría
\item \code{Año}: año de la serie
}
}
\source{
Instituto Provincial de Estadísticas y Censos (IPEC)
}
\usage{
data(tambos_2008_2015)
}
\description{
Datos sobre numero de establecimientos lecheros y composición de rodeo en la provincia de Santa Fe durante el período 2008-2015.
}
\keyword{internal}
