% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{tomate}
\alias{tomate}
\title{Crecimiento de tejido de tomate}
\format{
A data frame with 20 rows and 3 columns:
\itemize{
\item \code{Tejido}: id del tejido tratado.
\item \code{TR}: tratamiento aplicado (3\\% de fructosa, 3\\% de glucosa, 3\\% de
sacarosa y Control).
\item \code{CR}: crecimiento del tejido medido en mm X 10.
}
}
\source{
--
}
\usage{
data(tomate)
}
\description{
Experimento realizado para evaluar el crecimiento de secciones de tejido de
plantas de tomate en medios de cultivo con distintas cantidades y tipos de
azúcares.
}
\details{
Se evaluaron 4 tratamientos con 4 repeticiones cada uno:
\itemize{
\item 1- 3\% de fructosa.
\item 2- 3\% de glucosa.
\item 3- 3\% de sacarosa.
\item 4- Control (sin azúcar).
El resultado de crecimiento de tejidos (CR), se midió en mm X 10.
}
}
\keyword{internal}
