% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{tribolium}
\alias{tribolium}
\title{Pérdida de agua en Tribolium confusum}
\format{
A data frame with 9 rows and 2 columns:
\itemize{
\item \code{Humedad}: Humedad relativa (\%)
\item \code{Perdida}: Pérdida de peso (mg)
}
}
\usage{
data(tribolium)
}
\description{
Estudio de pérdida de agua en el coleóptero de la harina (\emph{Tribolium confusum}).
}
\details{
Se pesaron 9 lotes de 25 coleópteros cada uno, se guardaron a diferente humedad relativa y se pesaron nuevamente a los 6 días. Se calculó la pérdida de peso (mg) para cada lote.
}
\keyword{internal}
