% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{variedad_trigo}
\alias{variedad_trigo}
\title{Rendimiento de variedades de trigo}
\format{
A data frame with 16 rows and 4 columns:
\itemize{
\item \code{F}: id de la fila.
\item \code{C}: id de la columna.
\item \code{VAR}: variedad de trigo evaluada.
\item \code{REND}: rendimiento en kg por parcela.
}
}
\source{
--
}
\usage{
data(variedad_trigo)
}
\description{
Mediante un experimento se evaluó los rendimientos de distintas variedades
de trigo en kilogramos por parcela (cada parcela de 42 metros cuadrados).
}
\details{
Se evaluaron cuatro variedades en un DCL.
}
\keyword{internal}
