% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_score.R
\name{z_score}
\alias{z_score}
\title{Calcular z-scores para determinar Outliers}
\usage{
z_score(data = NULL, col = NULL)
}
\arguments{
\item{data}{Un data.frame o grouped_df que contiene la columna a analizar (por defecto NULL).}

\item{col}{Nombre de la columna (carácter o nombre sin comillas) de la cual se desea obtener el z-score (por defecto NULL).}
}
\value{
Retorna de manera invisible el data.frame (o grouped_df) recibido con una columna extra \code{z} que contiene los z-scores calculados.
}
\description{
Calcula los valores estandarizados (z-scores) para una columna de un data.frame,
agregando una columna \code{z} con los resultados e informando por consola la cantidad de valores
que superan 2 y 3 desviaciones estándar.
}
\examples{
library(estadisticaFcaUnl)
data(fertilizante_pastizal)

# Usando nombre de columna como cadena
z_score(fertilizante_pastizal, "P")

# Usando nombre de columna sin comillas
z_score(fertilizante_pastizal, P)

}
