% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.test.R
\name{z_values}
\alias{z_values}
\title{Internal function to calculate p-value and confidence interval}
\usage{
z_values(
  z_calc,
  x,
  se,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95
)
}
\arguments{
\item{z_calc}{Numeric. Calculated Z statistic.}

\item{x}{Numeric. Center of the confidence interval (mean or difference).}

\item{se}{Numeric. Standard error.}

\item{alternative}{Character string specifying the alternative hypothesis.
Must be one of "two.sided" (default), "greater" or "less".}

\item{conf.level}{Numeric. Confidence level of the interval. Default is 0.95.}
}
\value{
A list containing `p_val` (p-value) and `conf_int` (confidence interval).
}
\description{
Helper function to calculate p-values and confidence intervals for z-tests.
}
\keyword{internal}
