% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{ace_codornices}
\alias{ace_codornices}
\title{Acetil colinesterasa en codornices}
\format{
A data frame with 10 rows and 2 columns:
\describe{
\item{Momento:}{momento en que se expuso las codornices a las condiciones
ambientales (antes o despues).}
\item{ACE:}{nivel de acetil colinesterasa registrado.}
}
}
\source{
--
}
\usage{
data(ace_codornices)
}
\description{
La acetil colinesterasa (ACE) es un indicador de la actividad cerebral.
Las codornices son muy susceptibles a un determinado pesticida usado en la
producción de algodón ya que cantidades residuales de dicho pesticida produce
cambios en su actividad cerebral.
}
\details{
Para determinar si una nueva formulación disminuye la deriva de este
producto, se utilizaron dos grupos de codornices:
\describe{
\item{Grupo 1:}{seis codornices expuestas a las condiciones ambientales
antes de la aplicación.}
\item{Grupo 2:}{cuatro codornices expuestas a las condiciones ambientales
luego de la aplicación.}
}

En ambos grupos se midió los niveles de ACE.
}
\keyword{datasets}
