% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{actividad}
\alias{actividad}
\title{Actividad predominante de explotaciones agropecuarias}
\format{
A data frame with 12 rows and 3 columns:
\itemize{
\item \code{actividad}: actividad predominante de la explotación ("Agrícola", "Ganadera", "Mixta")
\item \code{localidad}: localidad del área rural ("A", "B", "C", "D")
\item \code{freq}: cantidad de explotaciones (frecuencia)
}
}
\usage{
data(actividad)
}
\description{
Frecuencias de la actividad económica predominante de explotaciones agropecuarias en el área rural de cuatro localidades.
}
\details{
En una encuesta socioeconómica efectuada en el área rural de cuatro localidades se obtuvieron los valores para la variable actividad predominante de la explotación (agrícola, ganadera, etc.).
}
\keyword{datasets}
