% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{agua_forrajera}
\alias{agua_forrajera}
\title{Relación entre agua aplicada y rendimiento en especie forrajera}
\format{
A data frame with 5 rows and 2 columns:
\itemize{
\item \code{agua}: cantidad de agua aplicada (cm^3)
\item \code{MS}: rendimiento de la especie forrajera (tn MS/ha)
}
}
\usage{
data(agua_forrajera)
}
\description{
Datos de cantidad de agua aplicada (en cm^3) y el rendimiento resultante de materia seca en una especie forrajera.
}
\details{
Los datos representan la cantidad de agua aplicada (cm^3) y el rendimiento de una especie forrajera (tn MS/ha), en un establecimiento de la provincia de Santa Fe.
}
\keyword{datasets}
