% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{almacenamiento_carne}
\alias{almacenamiento_carne}
\title{Técnicas de envasado de carne}
\format{
A data frame with 12 rows and 3 columns:
\itemize{
\item \code{Corte}: identificación del corte de carne.
\item \code{Tratamiento}: método de envasado (al vacío, CO2, comercial, mezcla de gases).
\item \code{bacterias}: recuento de bacterias en log10 UFC/cm2.
}
}
\source{
--
}
\usage{
data(almacenamiento_carne)
}
\description{
Estudio del efecto de técnicas de envasado de carne en la supresión del
crecimiento bacteriano de cortes en góndola.
}
\details{
Cortes de carne envasados con diferentes tratamientos (al vacío, CO2,
comercial y mezcla de gases) para evaluar el crecimiento bacteriano.
}
\keyword{datasets}
