% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{area_trigo}
\alias{area_trigo}
\title{Relación entre tamaño de explotación y área sembrada con trigo}
\format{
A data frame with 12 rows and 2 columns:
\itemize{
\item \code{sup_total}: superficie o tamaño total de la explotación agrícola
\item \code{sup_trigo}: superficie o área sembrada con trigo
}
}
\usage{
data(area_trigo)
}
\description{
Datos de una muestra de 12 explotaciones agrícolas que registran el tamaño total de la explotación y el área sembrada con trigo.
}
\details{
A continuación se detallan los datos de una muestra de 12 explotaciones agrícolas donde se registra el tamaño de las mismas y el área sembrada con trigo (ambas variables en hectáreas o unidades de superficie correspondientes).
}
\keyword{datasets}
