% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{calificaciones}
\alias{calificaciones}
\title{Calificaciones de alumnos en Física y Matemática}
\format{
A data frame with 12 rows and 2 columns:
\itemize{
\item \code{Fisica}: calificación obtenida en la asignatura de Física
\item \code{Matematica}: calificación obtenida en la asignatura de Matemática
}
}
\usage{
data(calificaciones)
}
\description{
Calificaciones obtenidas en las asignaturas de Física y Matemática en una muestra de 12 alumnos.
}
\details{
La tabla detalla las calificaciones en Física y Matemática de una muestra de 12 alumnos seleccionados al azar.
}
\keyword{datasets}
