% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{cana_azucar}
\alias{cana_azucar}
\title{Variedades de caña de azúcar}
\format{
A data frame with 25 rows and 4 columns:
\describe{
\item{F:}{id de la fila.}
\item{C:}{id de la columna.}
\item{VAR:}{variedad de caña de azúcar evaluada.}
\item{RTO:}{producción en kilos por parcela.}
}
}
\source{
--
}
\usage{
data(cana_azucar)
}
\description{
Se estudió la producción de distintas variedades de caña de azucar siguiendo
un diseño en cuadrado latino.
}
\details{
Resultados de producción en kilos por parcela.
}
\keyword{datasets}
