% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{cana_pd}
\alias{cana_pd}
\title{Fechas de plantación en caña de azúcar}
\format{
A data frame with 45 rows and 4 columns:
\describe{
  \item{Fecha:}{fecha de plantación (parcela principal).}
  \item{Método:}{método de plantación (subparcela).}
  \item{Bloque:}{id del bloque.}
  \item{Rend:}{rendimiento en Tn/ha de azúcar.}
}
}
\source{
--
}
\usage{
data(caña_pd)
}
\description{
Se midió el rendimiento en Tn/ha de Azúcar que experimentan tres fechas de
plantación y tres métodos para plantar.
}
\details{
Diseño en parcelas divididas en bloques completamente al azar. Las fechas se
asignaron aleatoriamente a las tres parcelas principales de cada bloque, y
los 3 métodos se asignaron aleatoriamente a las subparcelas.
}
\keyword{datasets}
