% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{cangrejo_ermitanio}
\alias{cangrejo_ermitanio}
\title{Recuento de cangrejos ermitaños}
\format{
A data frame with 150 rows and 2 columns:
\itemize{
\item \code{cant}: número de cangrejos ermitaños registrados.
\item \code{Sitio}: sitio costero evaluado (Factor con 6 niveles: "1", "2", "3", "4", "5", "6").
}
}
\source{
--
}
\usage{
data(cangrejo_ermitanio)
}
\description{
Recuento de cangrejos ermitaños en 6 sitios costeros.
}
\details{
Para el estudio, se seleccionaron 6 sitios costeros. En cada uno de ellos
se trazaron transectas y se registró y contabilizó la cantidad de cangrejos
ermitaños presentes en cada una de ellas.
}
\keyword{datasets}
