% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{capturados}
\alias{capturados}
\title{Captura de insectos en intervalos de tiempo}
\format{
A data frame with 6 rows and 2 columns:
\itemize{
\item \code{obs}: cantidad de intervalos observados con el número de insectos correspondiente (frecuencia)
\item \code{insectos}: cantidad de insectos capturados en el intervalo
}
}
\usage{
data(capturados)
}
\description{
Distribución de frecuencia del número de insectos capturados en un lote durante 1000 intervalos de un minuto.
}
\details{
El número observado de insectos capturados en un lote en 1000 intervalos de un minuto se describe en esta tabla de frecuencias.
}
\keyword{datasets}
