% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{cemento}
\alias{cemento}
\title{Calidad de mezclas de cemento de diferentes proveedores}
\format{
A data frame with 24 rows and 4 columns:
\itemize{
\item \code{resistencia}: resistencia registrada en pruebas de laboratorio (indicador de calidad).
\item \code{partida}: factor que identifica la partida o muestra (1 o 2).
\item \code{mezcla}: tipo de mezcla (A, B, C, D) anidada dentro de proveedor.
\item \code{proveedor}: factor que identifica al proveedor (1, 2, 3).
}
}
\source{
--
}
\usage{
data(cemento)
}
\description{
Ensayo diseñado para evaluar si existen diferencias de calidad del cemento
en 4 mezclas elaboradas por 3 proveedores.
}
\details{
La calidad del cemento se midió mediante pruebas de resistencia en laboratorio.
Las mezclas se denominaron A, B, C y D, pero sus componentes varían de
proveedor en proveedor, por lo tanto la mezcla es un factor anidado dentro
del proveedor. De cada partida se tomaron 2 muestras para las determinaciones
(identificadas en el factor \code{partida}).
}
\keyword{datasets}
