% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{cerdos}
\alias{cerdos}
\title{Ganancia de peso en cerdos según tipo de alimentación}
\format{
A data frame with 16 rows and 3 columns:
\itemize{
\item \code{cerdo}: identificador numérico del animal
\item \code{alimento}: tipo de alimentación suministrada ("alim_d" o "alim_m")
\item \code{ganancia}: ganancia en peso al mes de iniciado el ensayo (en kg)
}
}
\usage{
data(cerdos)
}
\description{
Datos de ganancia en peso (en kg) de cerdos sometidos a dos tipos de alimentación diferentes (D y M).
}
\details{
En un ensayo de engorde de cerdos se utilizaron dos tipos de alimentación diferentes, D y M, en dos lotes de 8 cerdos. Al mes de iniciado el ensayo se registró la ganancia en peso (kg) por animal.
}
\keyword{datasets}
