% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{cesped_compactacion}
\alias{cesped_compactacion}
\title{Durabilidad de variedades de césped bajo compactación}
\format{
A data frame with 54 rows and 4 columns:
\itemize{
\item \code{Localidad}: factor que identifica la localidad o bloque (1, 2, 3).
\item \code{Compact}: factor que identifica el nivel de compactación (Leve, Moderado, Nada).
\item \code{Variedad}: factor que identifica la variedad de césped (V1 a V6).
\item \code{MS}: cantidad de materia seca registrada en la subunidad.
}
}
\source{
--
}
\usage{
data(cesped_compactacion)
}
\description{
Experimento para evaluar la durabilidad de 6 variedades de céspedes sometidas
a tres niveles de compactación en un diseño en franjas (strip-plot).
}
\details{
El ensayo se llevó a cabo en tres localidades (bloques). En cada bloque, cada
variedad de césped fue ubicada en parcelas. Debido a la dificultad para
maniobrar la maquinaria de compactación, esta se aplicó en franjas
transversales sobre las variedades.

La variable de respuesta registrada fue la cantidad de materia seca (MS) en
cada subunidad.
}
\keyword{datasets}
