% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{cesped_golf}
\alias{cesped_golf}
\title{Calidad de césped en campo de golf bajo fertilización y acumulación de biomasa}
\format{
A data frame with 24 rows and 5 columns:
\itemize{
\item \code{Fuente}: factor que identifica la fuente de fertilizante nitrogenado (Isobutenildiurea, Sulfato de Amonio, Urea, Urea c/ cubierta de sulfuro).
\item \code{Bloque}: factor que identifica el bloque o réplica (1, 2).
\item \code{Años}: años de acumulación de biomasa muerta (2, 5, 8).
\item \code{clorofila}: contenido de clorofila en la muestra de césped (mg/g).
\item \code{Años_ord}: años de acumulación de biomasa transformados a factor ordenado (2 < 5 < 8).
}
}
\source{
--
}
\usage{
data(cesped_golf)
}
\description{
Experimento para determinar el efecto de cuatro fuentes de nitrógeno y tres
niveles de acumulación de biomasa muerta sobre la calidad del césped (medida
como contenido de clorofila) en un diseño de parcelas divididas (split-plot).
}
\details{
En un campo de golf sembrado con pasto Pencross se delimitaron 8 parcelas principales
a las cuales se les asignó una fuente de fertilizante siguiendo un diseño en
bloques completos al azar (DBCA) con dos bloques (réplicas). Las fuentes evaluadas
fueron: urea, sulfato de amonio, isobutildenediurea (IBDU) y urea con cubierta
de sulfuro (UCS).

Dos años más tarde, cada una de las parcelas se dividió en tres subparcelas a las
que se asignaron al azar los niveles de acumulación de biomasa: 2, 5 y 8 años.
Al final del experimento se midió el contenido de clorofila (mg/g) en muestras
de césped.
}
\keyword{datasets}
