% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{coleoptero}
\alias{coleoptero}
\title{Pérdida de agua en coleópteros de la harina}
\format{
A data frame with 9 rows and 2 columns:
\itemize{
\item \code{HR}: humedad relativa a la que fueron expuestos los coleópteros (\%)
\item \code{perdida}: pérdida de peso registrada en el lote (mg)
}
}
\usage{
data(coleoptero)
}
\description{
Datos de pérdida de peso (en mg) en lotes de coleópteros de la harina (\emph{Tribolium confusum}) sometidos a distintas humedades relativas.
}
\details{
Los datos proceden de un estudio de pérdida de agua en el coleóptero de la harina (\emph{Tribolium confusum}). Se pesaron 9 lotes de 25 coleópteros cada uno, se guardaron a diferente humedad relativa (HR) y se pesaron nuevamente a los 6 días. Se calculó la pérdida de peso (mg) para cada lote.
}
\keyword{datasets}
