% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{colesterol}
\alias{colesterol}
\title{Nivel de colesterol en sangre por tipo de dieta}
\format{
A data frame with 43 rows and 3 columns:
\itemize{
\item \code{paciente}: identificador numérico del paciente
\item \code{tratamiento}: tipo de dieta seguida ("baja_grasa" o "normal")
\item \code{colesterol}: nivel de colesterol medido en sangre (mg/dl)
}
}
\usage{
data(colesterol)
}
\description{
Datos del nivel de colesterol en sangre (mg/dl) de pacientes bajo dos tipos de dietas: baja en grasa y normal.
}
\details{
Se compara el nivel de colesterol en la sangre (mg/dl) de pacientes seleccionados al azar y sometidos a dos dietas distintas, una baja en grasa y otra normal.
}
\keyword{datasets}
