% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{competencia}
\alias{competencia}
\title{Competencia intraespecífica en cultivos}
\format{
A data frame with 15 rows and 3 columns:
\itemize{
\item \code{D}: densidad de plantas.
\item \code{rend}: rendimiento de granos por parcela.
\item \code{D_ord}: densidad de plantas transformada a factor ordenado (10 < 20 < 30 < 40 < 50).
}
}
\source{
--
}
\usage{
data(competencia)
}
\description{
En un experimento de cultivos se evaluó la relación entre la densidad de
plantas (competencia) y la producción de granos.
}
\details{
Se usaron 5 densidades de plantas (10, 20, 30, 40 y 50) las que se asignaron
por triplicado a parcelas experimentales siguiendo un diseño completamente
al azar (DCA). A cosecha se determinó el rendimiento por parcela.
}
\keyword{datasets}
