% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{conc_glucosa}
\alias{conc_glucosa}
\title{Concentración de glucosa}
\format{
A data frame with 18 rows and 4 columns:
\describe{
  \item{Metodo:}{método químico de determinación (1 o 2).}
  \item{Nivel:}{nivel de glucosa aplicado (A, B o C).}
  \item{Concentracion:}{concentración de glucosa medida.}
  \item{repeticion:}{id de la repetición.}
}
}
\source{
--
}
\usage{
data(Conc_glucosa)
}
\description{
Experimento para determinar el efecto de dos métodos químicos de determinación
de glucosa y 3 niveles de glucosa aplicados al suero.
}
\details{
Se aplicaron las combinaciones de método y nivel de glucosa por triplicado a
muestras de suero.
}
\keyword{datasets}
