% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{concentracion_insecticida}
\alias{concentracion_insecticida}
\title{Comparación de métodos de estimación de concentración de insecticida}
\format{
A data frame with 20 rows and 2 columns:
\itemize{
\item \code{metodo}: método de medición utilizado ("A" o "B")
\item \code{concentracion}: concentración estimada del principio activo
}
}
\usage{
data(concentracion_insecticida)
}
\description{
Datos de concentraciones medidas con dos métodos rápidos en 10 dosis preparadas en laboratorio, con el fin de comparar ambos métodos.
}
\details{
Se quiere comparar dos métodos rápidos para estimar la concentración de un principio activo en un insecticida. Se preparan 10 dosis en el laboratorio y se mide la concentración de cada una con ambos métodos (A y B).
}
\keyword{datasets}
