% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{conservante_frutilla}
\alias{conservante_frutilla}
\title{Conservantes en frutillas congeladas}
\format{
A data frame with 32 rows and 2 columns:
\describe{
  \item{Tratamiento:}{preservante utilizado.}
  \item{decoloracion:}{grado de decoloración en una escala del 1 al 10.}
}
}
\source{
--
}
\usage{
data(conservante_frutilla)
}
\description{
Investigadores del Departamento de Cultivos Frutales de una universidad
compararon cuatro preservantes para ser utilizados en frutilla congeladas.
}
\details{
Se prepararon en recipientes para freezer muestras de frutillas a las cuales
al azar les agregaron uno de los 4 preservantes evaluados: uno estandar o
control y 3 nuevos compuestos.

Luego del tratamiento con el preservante, cada bandeja fue envuelta con bolsa
de plástico y conservada en el freezer a 0°C durante 6 meses. Luego de dicho
período las muestras fueron descongeladas y se realizó una evaluación de
decoloración en una escala de 1 a 10 (a mayor numero mayor decoloración).
}
\keyword{datasets}
