% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{corderos_des}
\alias{corderos_des}
\title{Efecto de DES en fosfolípidos de corderos}
\format{
A data frame with 20 rows and 3 columns:
\itemize{
\item \code{Momento}: momento del día en que se sacrificaron los corderos (AM o PM).
\item \code{DES}: tratamiento con dietilestilbestrol (no o si).
\item \code{FL}: contenido de fosfolípidos en sangre.
}
}
\source{
--
}
\usage{
data(corderos_des)
}
\description{
Experimento para determinar los efectos del dietilestilbestrol (DES) en el
contenido de fosfolípidos en sangre de corderos.
}
\details{
Se condujo un experimento para determinar los efectos del dietilestilbestrol
(DES) en el contenido de fosfolípidos en sangre de corderos sacrificados en
dos momentos del día (AM y PM). Para ello, se seleccionaron 20 corderos los
que se asignaron a los tratamientos por quintuplicado siguiendo un DCA.
}
\keyword{datasets}
