% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_groups.R
\docType{data}
\name{datasets_estb}
\alias{datasets_estb}
\title{Datasets para Estadística Básica}
\description{
Lista de datasets utilizados en la asignatura Estadística Básica.
}
\section{Datasets}{

\itemize{
\item \link{lisina_pollos}
\item \link{lotes_BA}
\item \link{pesada_novillos}
\item \link{pesada_terneros}
\item \link{productores}
\item \link{riesgo_arbolado}
\item \link{ryegrass}
\item \link{superficies_EAP}
\item \link{tambos_2008_2015}
\item \link{tribolium}
}
}

