% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{densidad_trigo_riego_suelo}
\alias{densidad_trigo_riego_suelo}
\title{Densidad de siembra en trigo bajo riego y gradientes de suelo}
\format{
A data frame with 25 rows and 6 columns:
\itemize{
\item \code{Fila}: factor de fila en el diseño DCL (1 a 5).
\item \code{Columna}: factor de columna en el diseño DCL (1 a 5).
\item \code{Letra}: letra asignada en el cuadrado latino (A a E).
\item \code{Densidad}: densidad de siembra (lb/acre).
\item \code{rend}: rendimiento registrado (lb/acre x 10^2^).
\item \code{Densidad_ord}: densidad de siembra transformada a factor ordenado (30 < 80 < 130 < 180 < 230).
}
}
\source{
--
}
\usage{
data(densidad_trigo_riego_suelo)
}
\description{
Se realizó un ensayo a campo para determinar la relación entre el
rendimiento de trigo y la densidad de siembra.
}
\details{
Se evaluaron 5 densidades de siembra (30, 80, 130, 180 y 230 lb/acre) en un
campo regado que tenía un gradiente de humedad en una dirección y un gradiente
de suelos en otra dirección. Para filtrar estos efectos se utilizó un diseño
de cuadrado latino (DCL).

En cada parcela el rendimiento fue registrado en lb/acre x 10^2^.
}
\keyword{datasets}
