% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{dieta_bufalos}
\alias{dieta_bufalos}
\title{Pérdida de peso en búfalos bajo dieta hipocalórica}
\format{
A data frame with 20 rows and 3 columns:
\itemize{
\item \code{animal}: identificador del búfalo
\item \code{tratamiento}: momento del registro del peso ("inicial" o "final")
\item \code{perdida}: peso del animal registrado (en kg, cuya diferencia representa la pérdida de peso)
}
}
\usage{
data(dieta_bufalos)
}
\description{
Datos de peso (en kg) medidos en 10 búfalos antes y después de ser sometidos a una dieta hipocalórica.
}
\details{
Se estudió el efecto en la pérdida de peso (kg) en 10 búfalos sometidos a una dieta hipocalórica. Los datos registran el peso de cada animal en el momento inicial y en el momento final del tratamiento.
}
\keyword{datasets}
