% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{dieta_pavos}
\alias{dieta_pavos}
\title{Dietas en pavos}
\format{
A data frame with 30 rows and 2 columns:
\itemize{
\item \code{Dieta}: id de la dieta suministrada.
\item \code{GP}: ganancia de peso (GP) registrada.
}
}
\source{
--
}
\usage{
data(dieta_pavos)
}
\description{
Experimento realizado para probar los efectos de 5 dietas en pavos.
}
\details{
Se asignaron aleatoriamente 6 pavos a cada una de las 5 dietas y los
alimentaron individualmente durante un determinado período de tiempo. En el
experimento se utilizaron pavos con similar peso inicial.

Luego se midió la ganancia de peso (GP) de cada pavo.
}
\keyword{datasets}
