% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{dietas_novillos}
\alias{dietas_novillos}
\title{Dietas en novillos}
\format{
A data frame with 20 rows and 3 columns:
\describe{
\item{Dieta:}{tipo de dieta suministrada.}
\item{Animal:}{id del animal.}
\item{peso:}{aumento de peso en kg.}
}
}
\source{
--
}
\usage{
data(dietas_novillos)
}
\description{
Resultados de un ensayo de nutrición animal donde se evaluaron 4 dietas o
raciones para engorde de novillos.
}
\details{
Para el experimento se contó con 24 novillos de la mismo raza, sexo, y
similar edad similar peso inicial. Cada animal fue asignado de manera
aleatoria a una de las 4 raciones evaluadas.

Luego de un cierto período se midió el aumento de peso (en kg) de cada uno de
los animales.
}
\keyword{datasets}
