% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{dms_ensilaje}
\alias{dms_ensilaje}
\title{Digestibilidad de materia seca en ensilaje de maíz}
\format{
A data frame with 13 rows and 2 columns:
\describe{
  \item{Trat:}{tipo de especie a la que se le suministro la dieta (cordero o
  novillo).}
  \item{DMS:}{digestibilidad de la materia seca en \%.}
}
}
\source{
--
}
\usage{
data(dms_ensilaje)
}
\description{
Ensayo de digestibilidad in vivo realizado para determinar si la
digestibilidad de la materia seca (DMS) de ensilaje de maíz es similar en
corderos y novillos.
}
\details{
Se seleccionaron 7 corderos y 6 novillos a los cuales se les suministró el
ensilaje y se midió la digestibilidad de lo consumido.
}
\keyword{datasets}
