% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{educacion}
\alias{educacion}
\title{Permanencia de estudiantes universitarios por estado civil}
\format{
A data frame with 4 rows and 3 columns:
\itemize{
\item \code{momento}: momento del registro ("ingreso", "3er_año")
\item \code{casado}: estado civil casado ("si", "no")
\item \code{frec}: cantidad de estudiantes (frecuencia)
}
}
\usage{
data(educacion)
}
\description{
Frecuencias de estudiantes universitarios según su estado civil y su permanencia en el establecimiento educacional tres años más tarde del ingreso.
}
\details{
Los datos corresponden al número de estudiantes universitarios hombres, casados y solteros, que aún estaban en el establecimiento educacional tres años más tarde del ingreso. Diseñado para contrastar si la proporción difiere entre grupos.
}
\keyword{datasets}
