% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{efectividad_sembradora}
\alias{efectividad_sembradora}
\title{Comparación de rendimiento por tipo de máquina sembradora}
\format{
A data frame with 20 rows and 3 columns:
\itemize{
\item \code{parcelas}: identificador numérico de la parcela
\item \code{maquina}: tipo de sembradora utilizada ("nueva" o "vieja")
\item \code{rendimiento}: rendimiento del cultivo obtenido (en kg/parcela)
}
}
\usage{
data(efectividad_sembradora)
}
\description{
Datos de rendimiento (en kg por parcela) obtenidos al comparar una máquina sembradora nueva con una tradicional.
}
\details{
Para contrastar la efectividad del uso de una nueva máquina sembradora con respecto a una tradicional se seleccionaron 10 parcelas. Cada una de ellas se dividió en dos sectores, asignándose al azar a cada uno de los mismos las distintas máquinas, utilizando el mismo cultivo. El rendimiento en kg/parcela se detalla en este set de datos.
}
\keyword{datasets}
