% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{energia_abejas}
\alias{energia_abejas}
\title{Energía en abejas}
\format{
A data frame with 27 rows and 6 columns:
\itemize{
\item \code{Temp}: temperatura del ambiente en °C.
\item \code{Visc}: viscosidad del líquido (concentración de sacarosa en \\%).
\item \code{Energia}: energía gastada por las abejas en J/s.
\item \code{Temp_ord}: temperatura transformada a factor ordenado.
\item \code{Visc_ord}: viscosidad transformada a factor ordenado.
\item \code{repeticion}: id de la repetición.
}
}
\source{
--
}
\usage{
data(energia_abejas)
}
\description{
Experimento para determinar el efecto de la temperatura del ambiente y la
viscosidad del líquido en el consumo de energía de abejas melíferas al beber.
}
\details{
Los niveles de temperatura experimentados fueron 20, 30 y 40°C, y la
viscosidad del líquido se controló por las concentraciones de sacarosa, que
eran de 20, 40 y 60\% del total de sólidos disueltos.

El entomólogo registró la energía gastada por las abejas en J/s.
}
\keyword{datasets}
