% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_bloqueo.R
\name{er_bloqueo}
\alias{er_bloqueo}
\title{Eficiencia Relativa de un Diseño en Bloques Completos al Azar o Cuadrado Latino}
\usage{
er_bloqueo(formula, data, comparacion = "DCAvsDBCA")
}
\arguments{
\item{formula}{Una fórmula del modelo del tipo \code{respuesta ~ bloque + tratamiento} (para DCA vs DBCA)
o \code{respuesta ~ fila + columna + tratamiento} (para DBCA vs DCL).}

\item{data}{Un data.frame que contiene las variables del modelo.}

\item{comparacion}{Carácter. Indica la comparación a realizar:
\code{"DCAvsDBCA"} (por defecto) para comparar DCA frente a DBCA, o
\code{"DBCAvsDCL"} para comparar DBCA frente a DCL (diseño en cuadrado latino).}
}
\value{
Retorna de manera invisible un valor numérico que representa la eficiencia relativa.
Para \code{"DBCAvsDCL"}, retorna de manera invisible un vector con dos valores correspondientes
a la eficiencia relativa considerando las filas como bloques y las columnas como bloques respectivamente.
}
\description{
Calcula la eficiencia relativa (ER) de aplicar un diseño de bloques
completos al azar (DBCA) frente a un diseño completamente al azar (DCA), o de un
diseño en cuadrado latino (DCL) frente a un diseño en bloques completos al azar (DBCA).
}
\details{
La función estima la eficiencia relativa basándose en la comparación de las
varianzas estimadas (CME) bajo los diseños involucrados y realiza una corrección por
grados de libertad. En la comparación \code{"DCAvsDBCA"}, el término del bloque debe especificarse
en primer lugar a la derecha de la fórmula. En la comparación \code{"DBCAvsDCL"}, los dos primeros
términos a la derecha de la fórmula representan las filas y las columnas respectivamente.
}
\examples{
library(estadisticaFcaUnl)

# Comparación DCA vs DBCA
data(fertilizante_pastizal)
er_bloqueo(P ~ Bloque + Tratamiento, fertilizante_pastizal)

# Comparación DBCA vs DCL
data(densidad_trigo_riego_suelo)
er_bloqueo(rend ~ Fila + Columna + Densidad_ord, densidad_trigo_riego_suelo, comparacion = "DBCAvsDCL")

}
