% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estadisticaFcaUnl-package.R
\docType{package}
\name{estadisticaFcaUnl-package}
\alias{estadisticaFcaUnl}
\alias{estadisticaFcaUnl-package}
\title{Herramientas y Datos para Estadística en FCA-UNL}
\description{
\if{html}{\figure{logo.png}{options: style='float: right;' alt='logo' width='120'}}

Provee conjuntos de datos y funciones auxiliares para la enseñanza
y práctica de la estadística en la Facultad de Ciencias Agrarias de la
Universidad Nacional del Litoral.
}
\section{Funciones}{

\itemize{
\item \link{vallas_outliers}: Detección de outliers.
\item \link{z.test}: Test Z.
\item \link{verificar_ortogonalidad}: Verificar ortogonalidad de matrices de contraste.
}
}

\section{Datasets}{

\itemize{
\item \link{datasets_estb}: Datasets para Estadística Básica.
\item \link{datasets_estex}: Datasets para Estadística Experimental.
}
}

\section{Funciones extra}{

\itemize{
\item \link{install.local_repo}: Instalación desde repositorio local.
\item \link{mkminiCRAN}: Creación de repositorio local.
}
}

\author{
\strong{Maintainer}: Alejandro Aléman \email{aaleman@fca.unl.edu.ar} (Cátedra de Estadística FCA-UNL)

Authors:
\itemize{
  \item Patricia M. Acetta \email{pacetta@fca.unl.edu.ar} (Cátedra de Estadística FCA-UNL)
  \item Carlos Agustin Alesso \email{calesso@fca.unl.edu.ar} (Cátedra de Estadística FCA-UNL)
}

}
\keyword{internal}
