% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{fert_trigo_bajo_riego}
\alias{fert_trigo_bajo_riego}
\title{Fertilización de trigo bajo riego}
\format{
A data frame with 24 rows and 3 columns:
\itemize{
\item \code{Bloque}: factor que identifica el bloque (1, 2, 3, 4).
\item \code{Tratamiento}: factor que identifica la estrategia de fertilización (1 a 6).
\item \code{N}: contenido de nitrógeno en el pseudo-tallo (ppm x 10^-1^).
}
}
\source{
--
}
\usage{
data(fert_trigo_bajo_riego)
}
\description{
Se evaluaron 6 estrategias (momento y dosis) de fertilización nitrogenada en
trigo, incluyendo un control sin fertilización, y se midió la respuesta como
contenido de N en el pseudo-tallo.
}
\details{
El ensayo se realizó en un campo con sistema de riego que producía un
gradiente de humedad. Para controlarlo, las parcelas se ubicaron en bloques
perpendiculares a dicho gradiente.

Los datos de nitrógeno se registran en ppm x 10^-1^.
}
\keyword{datasets}
