% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{fertilizante_pastizal}
\alias{fertilizante_pastizal}
\title{Fertilización en \emph{Bouteloua gracilis}}
\format{
A data frame with 25 rows and 3 columns:
\itemize{
\item \code{Bloque}: id del bloque.
\item \code{Tratamiento}: tipo de tratamiento aplicado.
\item \code{P}: concentración de P en tejido foliar.
}
}
\source{
--
}
\usage{
data(fertilizante_pastizal)
}
\description{
Ensayo de fertilización en un pastizal de \emph{Bouteloua gracilis} siguiendo un
diseño en bloques completos al azar.
}
\details{
Se evaluaron cinco tratamientos de fertilización (combinaciones de N y P2O5
más un control) los cuales fueron asignados a parcelas en cada bloque.

Como respuesta se cuantificó la concentración de P en tejido foliar.

Los tratamientos:
\itemize{
\item 1- control: sin fertilizar.
\item 2- N50: 50 lb de N.
\item 3- N100: 100 lb de N.
\item 4- N50P75: 50 lb de N + 75 lb de P~2~O~5~.
\item 5- N100P75: 100 lb de N + 75 lb de P~2~O~5~.
}
}
\keyword{datasets}
