% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{fibra_quimico}
\alias{fibra_quimico}
\title{Compuestos químicos en fibra}
\format{
A data frame with 20 rows and 3 columns:
\describe{
  \item{BL:}{id del bloque (muestra de fibra).}
  \item{CQ:}{tipo de compuesto químico aplicado.}
  \item{RF:}{resistencia de la fibra.}
}
}
\source{
--
}
\usage{
data(fibra_quimico)
}
\description{
Experimento realizado para determinar el efecto de cuatro compuestos químicos
diferentes en la resistencia de una fibra.
}
\details{
Se seleccionaron cinco muestras de fibra, y un diseño de bloque aleatorio. Se
ejecutó probando cada tipo de compuesto químico en orden aleatorio en cada
muestra de fibra.
}
\keyword{datasets}
