% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{fijacion_N}
\alias{fijacion_N}
\title{Fijación de Nitrógeno}
\format{
A data frame with 48 rows and 5 columns:
\itemize{
\item \code{Cultivo}: tipo de cultivo hospedante.
\item \code{N}: nivel de nitrógeno aplicado en ppm.
\item \code{acetileno}: reducción de acetileno medida.
\item \code{N_ord}: nivel de nitrógeno transformado a factor ordenado.
\item \code{repeticion}: id de la repetición.
}
}
\source{
--
}
\usage{
data(fijacion_N)
}
\description{
Experimento para determinar el efecto del N y hospedantes (cultivos) en la
fijación biológica de N de bacterias Rizhobium.
}
\details{
Se evaluaron cuatro cultivos (alfalfa, soya, guar y habas) y tres niveles de
N (0, 50 y 100 ppm N). Dos plantas de cada cultivo fueron inoculadas con
Rhizobium y cultivadas en frascos de Leonard con una de las tres
concentraciones de N.

La reducción de acetileno refleja la cantidad de nitrógeno que fija la
bacteria en la relación simbiótica con la planta (a mayor reducción, mayor
actividad).
}
\keyword{datasets}
