% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{forestal_alt}
\alias{forestal_alt}
\title{Altura de árboles en ensayo forestal en San Agustín}
\format{
A data frame with 40 rows and 4 columns:
\itemize{
\item \code{Bloque}: factor que identifica el bloque o repetición (1 a 4).
\item \code{Material}: factor que identifica el material genético del eucalipto.
\item \code{Manejo}: factor que identifica el tipo de manejo del suelo (C o T).
\item \code{altura}: altura del árbol al año de vida.
}
}
\source{
--
}
\usage{
data(forestal_alt)
}
\description{
Al año de vida de una plantación forestal experimental en San Agustín, se midió
la altura de árboles de cinco materiales genéticos bajo dos manejos de suelo.
}
\details{
Los cinco materiales genéticos evaluados fueron:
\itemize{
\item \verb{E. dunnii} (de semillas)
\item \code{GC08} (clon de \emph{E. grandis} x \emph{E. camaldulensis})
\item \code{GC09} (clon de \emph{E. grandis} x \emph{E. camaldulensis})
\item \code{H105} (clon de \emph{E. grandis} x \emph{E. camaldulensis})
\item \verb{GT H117} (clon de \emph{E. grandis} x \emph{E. tereticornis})
}

Los manejos de suelo evaluados fueron:
\itemize{
\item \code{C}: Control, sin subsolado.
\item \code{T}: Tratado, con subsolado a 60 cm de profundidad.
}
}
\keyword{datasets}
