% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{fungicida}
\alias{fungicida}
\title{Efectividad de dosis de fungicida para el control de roya en trigo}
\format{
A data frame with 13 rows and 2 columns:
\itemize{
\item \code{Dosis}: dosis de principio activo del fungicida aplicada (gr.p.a./ha)
\item \code{Daño}: daño evaluado como tamaño promedio de las manchas en hoja bandera
}
}
\usage{
data(fungicida)
}
\description{
Datos del daño remanente (tamaño promedio de manchas en hoja bandera) obtenido al evaluar distintas dosis de un nuevo fungicida contra la roya en trigo.
}
\details{
Se desea probar la efectividad de un nuevo fungicida para el control de roya en trigo. Se probaron distintas dosis en gramos de principio activo por hectárea (gr.p.a./ha) en 13 parcelas de 100 plantas cada una. A los 15 días de la aplicación se realizó una evaluación del daño, registrada como el tamaño promedio de las manchas en la hoja bandera.
}
\keyword{datasets}
