% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{germinacion}
\alias{germinacion}
\title{Pruebas de germinación con tratamientos químicos}
\format{
A data frame with 6 rows and 3 columns:
\itemize{
\item \code{Trat}: tratamiento químico aplicado ("A", "B", "C")
\item \code{Germinado}: estado de germinación de la semilla ("si", "no")
\item \code{semillas}: cantidad de semillas (frecuencia)
}
}
\usage{
data(germinacion)
}
\description{
Datos sobre los resultados de germinación en semillas sometidas a tres tratamientos químicos diferentes.
}
\details{
Un experimentador efectuó pruebas de germinación de tres tratamientos químicos aplicados a cada una de 200 semillas elegidas al azar.
}
\keyword{datasets}
