% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{gladiolos}
\alias{gladiolos}
\title{Gladiolos y herbicidas}
\format{
A data frame with 16 rows and 3 columns:
\itemize{
\item \code{BL}: id del bloque.
\item \code{TR}: tipo de tratamiento aplicado.
\item \code{EH}: peso promedio por inflorescencia en onzas.
}
}
\source{
--
}
\usage{
data(gladiolos)
}
\description{
Se compar칩 el efecto de varios herbicidas sobre el peso de las flores de
gladiolos.
}
\details{
El experimento se realiz칩 en un dise침o de bloques completos al azar con cuatro tratamientos y cuatro bloques.
\itemize{
\item 1- control (sin herbicida).
\item 2- 2,4D-TCA.
\item 3- DN/CR.
\item 4- SESIN.
Luego se midi칩 el peso promedio por inflorescencia en onzas.
}
}
\keyword{datasets}
