% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{guayule}
\alias{guayule}
\title{Contenido de caucho en guayule}
\format{
A data frame with 27 rows and 2 columns:
\describe{
\item{Tipo:}{tipo de ejemplar (atípicos o aberrantes).}
\item{caucho:}{contenido de caucho.}
}
}
\source{
--
}
\usage{
data(guayule)
}
\description{
El guayule (\emph{Parthenium argentatum}) es una planta arbustiva que se utiliza
como fuente alternativa de látex hipoalergénico.
}
\details{
En un estudio para caracterizar los contenidos de caucho se recolectaron 54
ejemplares los cuales se clasificaron en \code{atípicos} y \code{aberrantes}.

A cada individuo de estos grupos se le midió el contenido de caucho.
}
\keyword{datasets}
