% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{hibridos_maiz}
\alias{hibridos_maiz}
\title{Híbridos de maíz}
\format{
A data frame with 32 rows and 2 columns:
\describe{
\item{Hibrido:}{id del híbrido de maíz evaluado.}
\item{rend:}{rendimiento registrado.}
}
}
\source{
--
}
\usage{
data(hibridos_maiz)
}
\description{
Ensayo comparativo de rendimiento de maíz.
}
\details{
Se evaluaron 4 híbridos: A, B, C y D los cuales fueron asignados
aleatoriamente a cada una de las 32 parcelas disponibles para el estudio.
}
\keyword{datasets}
