% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{hileras_maiz}
\alias{hileras_maiz}
\title{Relación entre peso de espiga y número de hileras en maíz}
\format{
A data frame with 18 rows and 2 columns:
\itemize{
\item \code{Hileras}: número de hileras de grano en la espiga
\item \code{Peso}: peso de la espiga
}
}
\usage{
data(hileras_maiz)
}
\description{
Datos de peso de la espiga y el número de hileras de granos correspondientes, medidos en un estudio sobre maíz.
}
\details{
Un estudio sobre maíz busca determinar la relación entre el peso de la espiga (en gramos u otra unidad de peso) y el número de hileras de grano en la misma.
}
\keyword{datasets}
