% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{huevos}
\alias{huevos}
\title{Relación entre peso y producción de huevos en langostas}
\format{
A data frame with 30 rows and 2 columns:
\itemize{
\item \code{Huevos}: número de huevos producidos por la hembra
\item \code{Peso}: peso de la hembra de langosta (en gramos)
}
}
\usage{
data(huevos)
}
\description{
Datos sobre el peso de hembras de langostas y el número de huevos producidos.
}
\details{
En un estudio sobre el éxito reproductivo de langostas, un entomólogo busca ajustar y modelar la relación entre el peso de las hembras y el número de huevos que produce. Selecciona 30 langostas de diferentes pesos (en gramos) y registra el número de huevos producidos.
}
\keyword{datasets}
