% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{insecticida_poroto}
\alias{insecticida_poroto}
\title{Insecticidas en poroto}
\format{
A data frame with 12 rows and 3 columns:
\describe{
  \item{Bloque:}{id del bloque.}
  \item{Insecticida:}{id del insecticida aplicado.}
  \item{emergencia}{número de plántulas emergidas por fila.}
}
}
\source{
--
}
\usage{
data(insecticida_poroto)
}
\description{
Experimento realizado para comparar los efectos de tres insecticidas en una
variedad de frijoles (_Phaseolus vulgaris_).
}
\details{
Para el ensayo disponía de cuatro parcelas dentro del campo experimental.
Debido a que dichas parcelas tenían diferencias de fertilidad, drenaje y
protección a los vientos, el investigador decidió realizar el experimento
siguiendo un DBCA donde cada parcela fue considerada como un bloque.

A su vez, cada parcela se dividió en tres hileras distanciadas de modo tal
que el efecto de los insecticidas no se superponga.
En cada hilera se sembraron 100 semillas y los insecticidas se asignaron
aleatoriamente a cada fila dentro de la parcela.

Luego de un tiempo se registró el número de plántulas emergidas por fila.
}
\keyword{datasets}
