% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{invernadero_spartina}
\alias{invernadero_spartina}
\title{Crecimiento de \emph{Spartina altemiflora}}
\format{
A data frame with 48 rows and 3 columns:
\itemize{
\item \code{TR}: tipo de combinacion aplicada.
\item \code{BL}: id del bloque.
\item \code{BI}: bioamasa registrada en peso seco.
}
}
\source{
--
}
\usage{
data(invernadero_spartina)
}
\description{
Experimento realizado en invernadero sobre el crecimiento de
\emph{Spartina altemiflora} (gramínea perenne), planta ecológicamente importante
de las salinas, para evaluar los efectos de la salinidad, el nitrógeno y
aireación.
}
\details{
La variable respuesta es la biomasa, el peso seco de toda la parte aérea de
la planta.

Cada tratamiento (TR) son diferentes combinaciones de salinidad, el nitrógeno
y aireación, con un total de 12 tratamientos.
}
\keyword{datasets}
