% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{labores_remolacha}
\alias{labores_remolacha}
\title{Labores culturales en remolacha azucarera}
\format{
A data frame with 9 rows and 4 columns:
\describe{
  \item{F:}{id de la fila.}
  \item{C:}{id de la columna.}
  \item{LAB:}{tipo de labor cultural aplicada.}
  \item{RTO:}{rendimiento en toneladas por hectárea.}
}
}
\source{
--
}
\usage{
data(labores_remolacha)
}
\description{
Se midieron los rendimientos de remolacha azucarera en toneladas por
hectárea bajo tres tipos de labores culturales en un diseño en cuadrado
latino.
}
\details{
--
}
\keyword{datasets}
