% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{larvas_cesped}
\alias{larvas_cesped}
\title{Insecticidas en césped}
\format{
A data frame with 60 rows and 5 columns:
\describe{
\item{Sitio:}{id del sitio experimental.}
\item{Parcela:}{id de la parcela.}
\item{Tratamiento:}{id del insecticida aplicado.}
\item{Muestra:}{id de la muestra dentro de la parcela.}
\item{cantidad:}{número de larvas sobrevivientes.}
}
}
\source{
--
}
\usage{
data(larvas_cesped)
}
\description{
La larva del género \emph{Tipula sp.} se alimenta de raices y pelos radicales de
céspedes produciendo pérdida de plantas y amarillamiento.

Un laboratorio llevo a cabo un exeprimento para evaluar la efectividad de
cuatro alternativas de control químico.
}
\details{
Para ampliar el nivel de inferencia del estudio, los investigadores
seleccionaron seis sitios experimentales con distintas condiciones de suelo
dentro de los cuales asignaron de manera aleatoria los insecticidas a
parcelas sembradas con césped. Adicionalmente, en cada sitio una parcela fue
dejada sin tratar a modo de control.

A la semana siguiente se contó el número de larvas en tres sectores de cada
parcela y se registro el promedio de sobrevivientes de cada parcela.
}
\keyword{datasets}
