% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{lesiones_hojas}
\alias{lesiones_hojas}
\title{Lesiones virales en hojas de tabaco}
\format{
A data frame with 16 rows and 3 columns:
\itemize{
\item \code{id}: identificador numérico de la planta (par de hojas)
\item \code{preparado}: tipo de preparado del virus aplicado ("prep1" o "prep2")
\item \code{lesiones}: número de lesiones provocadas en la hoja
}
}
\usage{
data(lesiones_hojas)
}
\description{
Datos sobre la cantidad de lesiones provocadas artificialmente en hojas de tabaco por dos preparados con distintas concentraciones de un virus.
}
\details{
En un ensayo se comparó la cantidad de lesiones provocadas en hojas de tabaco por dos concentraciones de un virus. Se seleccionaron 8 pares de hojas (un par de hojas por planta). A cada hoja del par se le aplicó uno de los preparados (prep1 o prep2) y se midió el número de lesiones provocadas artificialmente.
}
\keyword{datasets}
