% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{lisina_pollos}
\alias{lisina_pollos}
\title{Suplementaci칩n con Lisina en pollos}
\format{
A data frame with 12 rows and 3 columns:
\itemize{
\item \code{Pollo}: Identificador del pollo
\item \code{Consumo}: Consumo de alimento
\item \code{Lisina}: Cantidad de lisina consumida (g)
}
}
\usage{
data(lisina_pollos)
}
\description{
Experimento para modelar la relaci칩n entre la ganancia de peso en pollos cuyas dietas fueron suplementadas con distintas cantidades de lisina.
}
\details{
Se seleccion칩 una muestra aleatoria de 12 pollos los cuales fueron alojados en jaulas individuales y alimentados a voluntad con el alimento m치s el suplemento.
}
\keyword{datasets}
