% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{lluvia_forraje}
\alias{lluvia_forraje}
\title{Rendimiento de forraje en relación a la lluvia caída}
\format{
A data frame with 10 rows and 2 columns:
\itemize{
\item \code{rend}: rendimiento de forraje obtenido (toneladas)
\item \code{lluvia}: cantidad de lluvia caída (mm)
}
}
\usage{
data(lluvia_forraje)
}
\description{
Datos de rendimiento de forraje y la lluvia caída durante los meses de julio y agosto en 10 establecimientos elegidos al azar.
}
\details{
Los datos representan el rendimiento en toneladas de forraje y la lluvia caída en mm durante los meses de julio y agosto en 10 establecimientos elegidos al azar en una determinada región.
}
\keyword{datasets}
