% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{lodos}
\alias{lodos}
\title{Acumulación de cinc en cebada fertilizada con lodos}
\format{
A data frame with 36 rows and 5 columns:
\itemize{
\item \code{Ciudad}: factor que identifica la ciudad de origen de los barros (A, B, C).
\item \code{Dosis}: dosis de barros aplicada (0.5, 1.0, 1.5).
\item \code{rep}: número de réplica o repetición (1 a 4).
\item \code{Zn}: contenido de cinc (Zn) en la planta.
\item \code{Dosis_ord}: dosis de barros transformada a factor ordenado (0.5 < 1.0 < 1.5).
}
}
\source{
--
}
\usage{
data(lodos)
}
\description{
Bioensayo para determinar la acumulación de cinc (Zn) en plantas de cebada
fertilizadas con barros (lodos) de plantas de tratamiento de efluentes
urbanos de tres ciudades.
}
\details{
Se evaluaron barros provenientes de tres ciudades (niveles A, B y C) aplicados
en tres dosis. Se midió el contenido de cinc (Zn) acumulado en la planta de
cebada.
}
\keyword{datasets}
