% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{met_pesados}
\alias{met_pesados}
\title{Metales pesados en lodos}
\format{
A data frame with 36 rows and 5 columns:
\itemize{
\item \code{Ciudad}: ciudad donde se realizó el ensayo.
\item \code{Dosis}: dosis de lodo aplicada.
\item \code{zinc}: nivel de zinc medido en las plantas.
\item \code{Dosis_ord}: dosis de lodo transformada a factor ordenado.
\item \code{repeticion}: id de la repetición.
}
}
\source{
--
}
\usage{
data(met_pesados)
}
\description{
Estudio sobre el efecto de lodos de desagüe en el crecimiento de plantas y
la absorción de metales pesados.
}
\details{
Los niveles de metales en los lodos se probaron según el crecimiento de
plantas en ambientes que contienen distintas dosis de lodo (bioensayos) en
diferentes ciudades.
}
\keyword{datasets}
