% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{molinos}
\alias{molinos}
\title{Resistencia al viento en torres de molinos}
\format{
A data frame with 15 rows and 1 column:
\itemize{
\item \code{km_h}: velocidad del viento (en km/h) soportada por la torre
}
}
\usage{
data(molinos)
}
\description{
Datos de resistencia al viento (en km/h) de una muestra aleatoria de 15 torres de molinos sometidas a pruebas de viento.
}
\details{
Una empresa mayorista solicita al fabricante torres de molinos que puedan soportar vientos de 80 \eqn{km~h^{-1}}. La misma quiere determinar si las torres se ajustan a estas especificaciones solicitadas, para lo cual selecciona una muestra aleatoria de 15 molinos y los somete a pruebas de viento.
}
\keyword{datasets}
