% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{MS_girasol}
\alias{MS_girasol}
\title{Porcentaje de materia seca en plantas de girasol}
\format{
A data frame with 30 rows and 3 columns:
\itemize{
\item \code{plantas}: identificador de la planta
\item \code{grupo}: clasificación de la planta ("Normales" o "Anormales")
\item \code{MS}: porcentaje de materia seca de la planta (\%)
}
}
\usage{
data(MS_girasol)
}
\description{
Datos del porcentaje de materia seca en plantas normales y anormales de un cultivo de girasol.
}
\details{
En un cultivo de girasol se seleccionaron 30 plantas al azar de las cuales 16 eran normales y 14 anormales. Los porcentajes de materia seca para cada grupo fueron registrados en este set de datos.
}
\keyword{datasets}
