% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{nectar_trebol_rojo}
\alias{nectar_trebol_rojo}
\title{Néctar en trébol rojo}
\format{
A data frame with 20 rows and 3 columns:
\describe{
\item{Cabeza:}{numero de identificación de la inflorescencia.}
\item{Presion:}{presión de vapor (4.4 o 9.9 mmHg) a la que fué sometida.}
\item{nectar:}{concentración de azúcares del néctar.}
}
}
\source{
--
}
\usage{
data(nectar_trebol_rojo)
}
\description{
Efecto de la presión de vapor en la secreción de néctar de flores de trébol
rojo (\emph{Trifolium pratense}).
}
\details{
Se tomó 10 cabezas (inflorescencias) de trébol rojo y las dividió a la mitad.
Cada mitad fue sometida a una presión de vapor (4.4 o 9.9 mmHg) y luego se
midió la concentración de azúcares del néctar.
}
\keyword{datasets}
