% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{neumaticos}
\alias{neumaticos}
\title{Comparación de duración de neumáticos de camiones}
\format{
A data frame with 30 rows and 3 columns:
\itemize{
\item \code{neumatico}: identificador numérico de la muestra de neumático
\item \code{marca}: marca/tipo del neumático ("A" o "B")
\item \code{duracion}: duración del neumático en miles de kilómetros (mil km)
}
}
\usage{
data(neumaticos)
}
\description{
Datos de duración (en miles de kilómetros) de dos marcas de neumáticos (A y B) probadas hasta el desgaste para camiones de una cooperativa agrícola.
}
\details{
Una cooperativa agrícola debe decidir cuál de dos tipos de neumáticos (A y B) va a comprar para sus camiones. Los neumáticos se prueban bajo condiciones semejantes hasta que se desgastan. Se emplean 15 de cada marca, registrándose la duración en miles de km.
}
\keyword{datasets}
