% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{ovejas}
\alias{ovejas}
\title{Ganancia en peso de ovejas}
\format{
A data frame with 25 rows and 1 column:
\itemize{
\item \code{peso}: ganancia en peso (en kilogramos)
}
}
\usage{
data(ovejas)
}
\description{
Datos de ganancia en peso (en kg) de una muestra de 25 ovejas sometidas a una ración alimenticia durante un cierto período.
}
\details{
La ganancia en peso de 25 ovejas sometidas a una ración alimenticia durante un cierto período fue registrada en este set de datos.
}
\keyword{datasets}
