% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{plantacion_dap}
\alias{plantacion_dap}
\title{Diámetro a la altura del pecho de ejemplares de Eucalyptus}
\format{
A data frame with 10 rows and 1 column:
\itemize{
\item \code{dap}: Diámetro a la altura del pecho en centímetros
}
}
\usage{
data(plantacion_dap)
}
\description{
Datos de DAP (diámetro a la altura del pecho) expresados en centímetros de los árboles de la plantación de \emph{Eucalyptus} de San Agustín (Santa Fe) en su segundo año.
}
\keyword{datasets}
