% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{pollos}
\alias{pollos}
\title{Suplementación con Lisina en pollos y ganancia de peso}
\format{
A data frame with 12 rows and 2 columns:
\itemize{
\item \code{ganancia}: ganancia de peso acumulada del pollo (en gramos)
\item \code{lisina}: cantidad estimada de lisina consumida por el pollo
}
}
\usage{
data(pollos)
}
\description{
Datos de consumo estimado de lisina y ganancia de peso en una muestra de 12 pollos.
}
\details{
Un investigador condujo un experimento para modelar la relación entre la ganancia de peso en pollos cuyas dietas fueron suplementadas con distintas cantidades de lisina (aminoácido) y la cantidad de lisina consumida. Dado que el porcentaje de lisina del alimento es conocido, la cantidad de lisina consumida se estima a partir de la cantidad de alimento consumida. Se seleccionó una muestra aleatoria de 12 pollos los cuales fueron alojados en jaulas individuales y alimentados a voluntad con el alimento más el suplemento.
}
\keyword{datasets}
