% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{productores}
\alias{productores}
\title{Producción de lotes por grupo de productores}
\format{
A data frame with 320 rows and 3 columns:
\itemize{
\item \code{Nivel_Produccion}: Nivel de producción del lote (Alto, Medio, Bajo)
\item \code{Grupo}: Grupo de productores (A, B, C)
\item \code{Lotes}: Cantidad de lotes (frecuencia)
}
}
\usage{
data(productores)
}
\description{
Datos de 320 lotes clasificados por producción de tres grupos de productores (A, B y C). Las clasificaciones se realizaron según el nivel de la producción de cada lote (Alto, Medio y Bajo).
}
\keyword{datasets}
