% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{proteina_caseina}
\alias{proteina_caseina}
\title{Contenido de proteína bruta y caseína en leche}
\format{
A data frame with 23 rows and 2 columns:
\itemize{
\item \code{PB}: porcentaje de proteína bruta en leche (\%)
\item \code{CA}: porcentaje de caseína en leche (\%)
}
}
\usage{
data(proteina_caseina)
}
\description{
Contenido de proteína bruta (PB) y caseína (CA) en leche registrado en una muestra de 23 tambos.
}
\details{
Los datos se refieren al contenido de proteína bruta (PB) y caseína (CA) en leche medido en una muestra de 23 tambos de la cuenca lechera del centro del país.
}
\keyword{datasets}
