% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{raciones_pollos}
\alias{raciones_pollos}
\title{Raciones alimenticias para pollos}
\format{
A data frame with 16 rows and 4 columns:
\describe{
  \item{Pisos:}{número de pisos de la jaula (filas).}
  \item{Casilleros:}{número de casilleros (columnas).}
  \item{Racion:}{ración alimenticia suministrada.}
  \item{Peso:}{peso del pollo en kg a las 8 semanas.}
}
}
\source{
--
}
\usage{
data(raciones_pollos)
}
\description{
Se probaron 4 raciones alimenticias para pollos, criados en jaula tipo batería de 4 pisos (filas) y 4 casilleros (columnas).
}
\details{
La variable analizada fué: peso del pollo (kg) a las 8 semanas de edad.
}
\keyword{datasets}
