% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{semilla_trigo}
\alias{semilla_trigo}
\title{Caracteres de semillas de trigo diploide}
\format{
A data frame with 190 rows and 6 columns:
\itemize{
\item \code{ID}: identificador numérico de la semilla
\item \code{Peso}: peso de la semilla
\item \code{Long}: longitud de la semilla
\item \code{Diametro}: diámetro de la semilla
\item \code{Humedad}: contenido de humedad de la semilla
\item \code{Dureza}: índice de dureza de la semilla
}
}
\usage{
data(semilla_trigo)
}
\description{
Datos de varios caracteres físicos (peso, longitud, diámetro, humedad, dureza) en una muestra de 190 semillas de trigo diploide (\emph{Triticum monococcum}).
}
\details{
Con el objeto de identificar variables asociadas con las diferencias de peso de semillas de trigo diploide (\emph{Triticum monococcum}), Jing et al. (2007) estudiaron varios caracteres (peso, longitud, diámetro, dureza) de 190 semillas obtenidas al azar.
}
\keyword{datasets}
