% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{semillas}
\alias{semillas}
\title{Proporción de tamaños de semilla de maíz en tres partidas}
\format{
A data frame with 9 rows and 3 columns:
\itemize{
\item \code{Partida}: partida de procedencia de la semilla ("A", "B", "C")
\item \code{Tamaño}: tamaño de la semilla ("grande", "normal", "chico")
\item \code{semillas}: cantidad de semillas (frecuencia)
}
}
\usage{
data(semillas)
}
\description{
Frecuencias de clasificación por tamaño de semillas de maíz tomadas de tres partidas diferentes.
}
\details{
Se quiere saber si la proporción de los tamaños de semilla de maíz son similares en 3 partidas de semillas. De cada partida se toma una muestra y se clasifica la semilla según su tamaño en chico, normal y grande.
}
\keyword{datasets}
