% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{semillas_trigo}
\alias{semillas_trigo}
\title{Rendimiento de semillas de trigo}
\format{
A data frame with 16 rows and 4 columns:
\describe{
  \item{Abono:}{tipo de abono utilizado (filas).}
  \item{Insecticida:}{tipo de insecticida utilizado (columnas).}
  \item{Semilla:}{tipo de semilla de trigo evaluada.}
  \item{Rendimiento:}{rendimiento registrado.}
}
}
\source{
--
}
\usage{
data(semillas_trigo)
}
\description{
Se realizó un ensayo para evaluar el rendimiento de la semilla de trigo.
}
\details{
Al plantear este experimento se pensó que podría conseguirse mayor precisión
si se controlaba la variabilidad introducida por los tipos de abono e
insecticida.
El instituto de experimentación agrícola está interesado en estudiar 4 tipos
de semilla de trigo, (A, B, C, D) y decide realizar el experimento
utilizando un diseño en cuadrado latino. Para ello selecciona 4 niveles para
cada una de las variables de bloque:
\itemize{
  \item Abono: (a1, a2, a3, a4)
  \item Insecticida: (i1, i2, i3, i4)
}
}
\keyword{datasets}
