% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{sorgo}
\alias{sorgo}
\title{Altura de plantas de sorgo forrajero}
\format{
A data frame with 16 rows and 1 column:
\itemize{
\item \code{altura}: altura de la planta (en metros)
}
}
\usage{
data(sorgo)
}
\description{
Datos de alturas de una muestra de 16 plantas de sorgo forrajero.
}
\details{
Las siguientes son las alturas de una muestra de 16 plantas de sorgo forrajero.
}
\keyword{datasets}
