% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{superficies_EAP}
\alias{superficies_EAP}
\title{Cantidad de empresas agropecuarias y la superficie promedio de la provincia de Santa Fe registradas en el Censo Nacional Agropecuario 2008.}
\format{
A data frame with 19 rows and 3 columns:
\describe{
  \item{Departamento}{nombre del departamento.}
  \item{Cant_EAP}{número de explotaciones agropecuarias.}
  \item{Sup_promedio}{superficie promedio de las explotaciones agropecuarias (en ha).}
}
}
\source{
Instituto Nacional de Estadísticas y Censos
}
\usage{
data(superficies_EAP)
}
\description{
Cantidad de empresas agropecuarias y la superficie promedio de la provincia de Santa Fe registradas en el Censo Nacional Agropecuario 2008.
}
\keyword{datasets}
