% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{tiempo_reaccion}
\alias{tiempo_reaccion}
\title{Tiempo de reacción en proceso químico}
\format{
A data frame with 25 rows and 5 columns:
\describe{
  \item{F:}{id de la fila.}
  \item{C:}{id de la columna.}
  \item{INGREDIENTE:}{ingrediente utilizado (A, B, C, D, E).}
  \item{TIEMPO:}{tiempo de reacción.}
}
}
\source{
--
}
\usage{
data(tiempo_reaccion)
}
\description{
Se evaluó el tiempo de reaccion de un proceso quimico con cinco ingredientes
diferentes en un diseño en cuadrado latino.
}
\details{
--
}
\keyword{datasets}
