% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estb.R
\docType{data}
\name{TRC}
\alias{TRC}
\title{Tasa de recuperación cardíaca (TRC)}
\format{
A data frame with 22 rows and 3 columns:
\itemize{
\item \code{individuo}: identificador numérico del participante
\item \code{Grupo}: grupo etario del participante ("1: 40-49 años", "2: 50-59 años")
\item \code{TRC}: tasa de recuperación cardíaca medida en tiempo (minutos y segundos, expresados decimalmente)
}
}
\usage{
data(TRC)
}
\description{
Datos de tasa de recuperación cardíaca (TRC) medida a través del tiempo registrado en una carrera de 2.5 km en dos grupos de participantes hombres.
}
\details{
Los participantes de un programa de salud miden su progreso mediante el tiempo que les toma correr determinada distancia. Un índice del progreso lo constituye la tasa de recuperación cardíaca (TRC). Los datos corresponden a tiempos (minutos y segundos) registrados para una carrera de 2.5 km. Los participantes pertenecen a dos grupos etarios: hombres entre 40 y 49 años (Grupo 1) y hombres entre 50 y 59 años (Grupo 2).
}
\keyword{datasets}
