% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{trebol_rizhobium}
\alias{trebol_rizhobium}
\title{Incoculación en trébol rojo}
\format{
A data frame with 30 rows and 2 columns:
\describe{
\item{Tratamiento:}{cepa con la que fué inoculada.}
\item{N:}{contenido de nitrogeno (en mg).}
}
}
\source{
--
}
\usage{
data(trebol_rizhobium)
}
\description{
Experimento realizado para evaluar la capacidad de fijación biológica de
nitrógeno de plantas de trébol rojo (\emph{Trifolium pratense}) inoculadas con
diferentes cepas de bacterias.
}
\details{
Se evaluaron tres tipos de tratamientos:
\describe{
\itemize{
\item 
\itemize{
\item 5 cepas de las bacterias \emph{Rizhobium trifolii}.
}
\item 
\itemize{
\item 5 cepas de las bacterias \emph{Rizhobium meliloti}.
}
\item 
\itemize{
\item combinación de 5 cepas de \emph{R. trifolii} con 5 cepas de \emph{R. meliloti}.
}
Los tratamientos fueron asignados por quintuplicado a macetas donde se dejó
crecer una planta de trébol rojo sobre la cual se midió el contenido de N (en mg).
}
}
}
\keyword{datasets}
