% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verificar_ortogonalidad.R
\name{verificar_ortogonalidad}
\alias{verificar_ortogonalidad}
\title{Verificar Ortogonalidad de una Matriz}
\usage{
verificar_ortogonalidad(M)
}
\arguments{
\item{M}{Una matriz numérica.}
}
\value{
Un valor lógico (invisible): TRUE si las columnas son ortogonales y suman cero, FALSE en caso contrario.
}
\description{
Comprueba si las columnas de una matriz son ortogonales entre sí y si suman cero (contrastes).
}
\examples{
K <- cbind(
    C1 = c(1, -1 / 4, -1 / 4, -1 / 4, -1 / 4),
    C2 = c(0, 1, -1, 0, 0),
    C3 = c(0, 0, 0, 1, -1),
    C4 = c(0, 1 / 2, 1 / 2, -1 / 2, -1 / 2)
)
verificar_ortogonalidad(K)
}
