% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_estex.R
\docType{data}
\name{vitA_vaquillas}
\alias{vitA_vaquillas}
\title{Vitamina A en vaquillas}
\format{
A data frame with 28 rows and 2 columns:
\itemize{
\item \code{Tratamiento}: tipo de dieta a la que fué sometida.
\item \code{GDPV}: ganancia diaria de peso vivo (GDPV) en g/día/animal.
}
}
\source{
--
}
\usage{
data(vitA_vaquillas)
}
\description{
Efecto de la Vitamina A en el crecimiento de vaquillas Holstein.
}
\details{
Se seleccionaron 28 vaquillas de similar peso inicial y edad, las cuales
fueron alimentadas individualmente durante un cierto período. A la mitad de
las vaquillonas se les suministró una dieta sin Vitamina A (control) y al
resto una dieta rica en Vitamina A. La asignación de las dietas fue
completamente al azar.

Al final del ensayo se registró el peso final y se calculó la ganancia
diaria de peso vivo (GDPV) en g/día/animal.
}
\keyword{datasets}
