% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.test.R
\name{z.test}
\alias{z.test}
\title{Normal Z test}
\usage{
z.test(
  x = NULL,
  y = NULL,
  sig_x,
  sig_y = NULL,
  alternative = c("two.sided", "less", "greater"),
  mu = 0,
  conf.level = 0.95,
  p1 = NULL,
  p2 = NULL,
  n1 = NULL,
  n2 = NULL
)
}
\arguments{
\item{x}{Numeric vector. Sample data for one-sample test or first sample for two-sample test.}

\item{y}{Numeric vector. Second sample data for two-sample test. Optional.}

\item{sig_x}{Numeric. Population standard deviation for x.}

\item{sig_y}{Numeric. Population standard deviation for y. Required for two-sample test.}

\item{alternative}{Character string specifying the alternative hypothesis.
Must be one of "two.sided" (default), "greater" or "less".}

\item{mu}{Numeric. Hypothesized true mean or difference in means. Default is 0.}

\item{conf.level}{Numeric. Confidence level of the interval. Default is 0.95.}

\item{p1}{Numeric. Proportion of success in population 1 (for two-proportion test).}

\item{p2}{Numeric. Proportion of success in population 2 (for two-proportion test).}

\item{n1}{Numeric. Sample size for population 1 (for two-proportion test).}

\item{n2}{Numeric. Sample size for population 2 (for two-proportion test).}
}
\value{
A list with class "htest" containing the following components:
  \item{statistic}{The value of the z-statistic.}
  \item{p.value}{The p-value for the test.}
  \item{conf.int}{A confidence interval for the mean or difference in means appropriate to the specified alternative hypothesis.}
  \item{estimate}{The estimated mean or difference in means/proportions.}
  \item{null.value}{The specified hypothesized value of the mean or mean difference.}
  \item{alternative}{A character string describing the alternative hypothesis.}
  \item{method}{A character string indicating what type of test was performed.}
  \item{data.name}{A character string giving the name(s) of the data.}
}
\description{
Performs one and two sample z-tests on vectors of data.
}
