% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.test.R
\name{z.test}
\alias{z.test}
\title{Test de distribucion Normal Z}
\usage{
z.test(
  x = NULL,
  y = NULL,
  sig_x,
  sig_y = NULL,
  alternative = c("two.sided", "less", "greater"),
  mu = 0,
  conf.level = 0.95,
  p1 = NULL,
  p2 = NULL,
  n1 = NULL,
  n2 = NULL
)
}
\arguments{
\item{x}{Vector numerico. Muestra para una muestra o primer muestra para dos muestras.}

\item{y}{Vector numerico. Segunda muestra para dos muestras. Optional.}

\item{sig_x}{Numerico. Desviacion estandar poblacional para x.}

\item{sig_y}{Numerico. Desviacion estandar poblacional para y. Requerido para dos muestras.}

\item{alternative}{String de caracteres que especifica la hipotesis alternativa.
Debe ser "two.sided" (default), "greater" o "less".}

\item{mu}{Numerico. Hipotesis nula para la media o diferencia de medias. Por defecto es 0.}

\item{conf.level}{Numerico. Nivel de confianza del intervalo. Por defecto es 0.95.}

\item{p1}{Numerico. Proporcion de exito en la poblacion 1 (para prueba de proporciones).}

\item{p2}{Numerico. Proporcion de exito en la poblacion 2 (para prueba de proporciones).}

\item{n1}{Numerico. Tamaño de la muestra para la poblacion 1 (para prueba de proporciones).}

\item{n2}{Numerico. Tamaño de la muestra para la poblacion 2 (para prueba de proporciones).}
}
\value{
Un objeto de la clase "htest" que contiene los siguientes componentes:
\itemize{
\item \code{statistic}: El valor del estadistico Z.
\item \code{p.value}: El p-valor para la prueba.
\item \code{conf.int}: Un intervalo de confianza para la media o diferencia en medias apropiado a la hipotesis alternativa especificada.
\item \code{estimate}: La estimacion de la media o diferencia en medias/proportiones.
\item \code{null.value}: El valor hipotetico de la media o diferencia en medias.
\item \code{alternative}: Un string descriptivo de la hipotesis alternativa.
\item \code{method}: Un string indicando el tipo de prueba realizada.
\item \code{data.name}: Un string que da el nombre(s) de los datos.
}
}
\description{
Realiza pruebas de hipotesis de una o dos muestras con distribucion normal.
}
\examples{
# One-sample Z-test
x <- rnorm(100, mean = 10, sd = 2)
z.test(x, sig_x = 2, mu = 10)

# Two-sample Z-test
y <- rnorm(100, mean = 12, sd = 2)
z.test(x, y, sig_x = 2, sig_y = 2)

# Two-proportion Z-test
z.test(p1 = 0.5, p2 = 0.6, n1 = 100, n2 = 100)
}
