% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.test.R
\name{z_values}
\alias{z_values}
\title{Funcion interna para calcular p-valor y intervalo de confianza}
\usage{
z_values(
  z_calc,
  x,
  se,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95
)
}
\arguments{
\item{z_calc}{Numerico. Estadistico Z calculado.}

\item{x}{Numerico. Centro del intervalo de confianza (media o diferencia).}

\item{se}{Numerico. Error estandar.}

\item{alternative}{String de caracteres que especifica la hipotesis alternativa.
Debe ser "two.sided" (default), "greater" o "less".}

\item{conf.level}{Numerico. Nivel de confianza del intervalo. Por defecto es 0.95.}
}
\value{
Un listado con \code{p_val} (valor p) y \code{conf_int} (intervalo de confianza).
}
\description{
Funcion auxiliar para calcular p-valor y intervalo de confianza para pruebas de distribucion normal Z.
}
\keyword{internal}
